### 2025-05-07 #242 @rolypolytoy
Picht
Submitting Author: (@rolypolytoy)
Package Name: Picht
One-Line Description of Package: Electron and ion optics simulation using the Finite Difference Method (FDM)
Repository Link:  https://github.com/rolypolytoy/picht/
Version submitted: v2.2.0
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15376240.svg)](https://doi.org/10.5281/zenodo.15376240)
Version accepted: TBD

---

Simulates electron and ion trajectories using the finite difference method. Exists as an easy, powerful alternative to commercial electrodynamics tools, and stands apart from existing open-source tools by using the finite difference method instead of the boundary element/finite element method, for better support in Dirichlet boundaries (grounded boundary conditions vs infinite). Uses the more physically accurate Lorentz force equation rather than the paraxial ray equation, computes relativistic corrections using the gamma factor, and calculates electron and ion trajectories through electrostatic lenses like cylindrical lenses and unipotential lenses. Example cases demonstrate physically realistic behaviors of electrons inside einzel lenses and through multi-lens systems, where Picht demonstrates simulations of complex focusing and defocusing behaviors, as well as crossovers, spherical aberration, and spot-size demagnification.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

**Data Processing/Munging:** Picht allows the initialization of custom electrode geometries and then generates electron trajectory data based on this. This allows a physically accurate analysis of the impact of variable geometries and voltages on charged particle beams of different kinds.

**Scientific Software Wrappers:** The code is modular enough to be used as a more general electrodynamics tool due to the clean nature of how the ODE solver can integrate with other functions, so I thought to put this in here too, since using scipy's ODE solvers for the niche application of numerical electrodynamics is novel enough to add to the scientific community, but derivative enough to require this category.

  Researchers, startups, and students in electrodynamics, electron optics, electron microscopy, or nanofabrication. It can be used to simulate or prototype electrostatic lens systems, scanning electron microscopes using predominantly electrostatic lenses, focused ion beam systems, and some varieties of mass spectrometers. It is also useful as an educational tool in the field of electron optics because of how intuitive the syntax is and how intentionally constrained the features are. Its power is in the large amount of electrical and geometric parameterization it allows, rather than a large amount of prebuilt functions.

 There exists a similar package called Traceon (https://github.com/leon-vv/traceon) but it's distinct enough that it supports a different use case. Traceon assumes the user is fluent in meshing, has more support for 3D particle tracing than the axisymmetric view, and has a larger learning curve. Picht uses no approximations but instead uses the Lorentz force equation for guaranteeably accurate physical behavior, uses better boundary conditions, has much simpler dependency management, and is intentionally constrained in scope to enable rapid prototyping and easy syntax. In addition- commercial software packages like ANSYS Maxwell or COMSOL, as well as most open-source electrodynamics software like Traceon use Neumann boundary conditions natively rather than Dirichlet boundary methods- and the differences in boundary effects are distinct enough that this is an important point of clarification. Dirichlet boundary conditions more accurately simulate finite container conditions- where the electrodes are surrounded by metal grounded containers rather than in an infinitely open space or with insulators surrounding.
