### 2025-05-07 #242 @coatless
---
---

Hi @rolypolytoy, thanks for submitting the package.

### Installation hiccups

I'm running into issues installing it as you're using a newer version of NumPy v2.2.z with `numpy.rec`; however, there is a downstream dependency of `thinc`, `mendeleev`, and `numba` that are problematic:

```
thinc 8.3.6 requires numpy<3.0.0,>=2.0.0, but you have numpy 1.26.4 which is incompatible.
mendeleev 1.0.0 requires numpy<2.0,>=1.21; python_version < "3.12", but you have numpy 2.2.5 which is incompatible.
numba 0.60.0 requires numpy<2.1,>=1.22, but you have numpy 2.0.2 which is incompatible.
```

Test document here:

https://github.com/coatless/pyopensci-eic/blob/main/i242-picht.ipynb

The package is missing a user-facing documentation website. Please see our guide here for help on creating docs using `sphinx`.

https://www.pyopensci.org/python-package-guide/documentation/index.html

## Actions

Please address the above points and ping me to restart the EIC checklist once complete.
