+++ picht.egg-info/SOURCES.txt
@@ -6,9 +6,9 @@ setup.py
 picht/__init__.py
 picht/core.py
 picht/example.py
-picht/unit_test.py
 picht.egg-info/PKG-INFO
 picht.egg-info/SOURCES.txt
 picht.egg-info/dependency_links.txt
 picht.egg-info/requires.txt
 picht.egg-info/top_level.txt
+tests/test_picht.py
