+++ pyproject.toml
@@ -16,10 +16,10 @@ authors = [
 
 dependencies = [
     "numpy==1.26.4",
-    "matplotlib==3.10.1",
-    "scipy==1.15.2",
-    "numba==0.61.2",
-    "mendeleev==1.0.0",
+    "matplotlib",
+    "scipy",
+    "numba",
+    "mendeleev",
 ]
 
 classifiers = [
