+++ picht.egg-info/requires.txt
@@ -1,5 +1,5 @@
-numpy==2.2.5
-matplotlib==3.10.1
-scipy==1.15.2
-numba==0.61.2
-mendeleev==1.0.0
+numpy==1.26.4
+matplotlib
+scipy
+numba
+mendeleev
