+++ pyproject.toml
@@ -1,10 +1,10 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel"]
+requires = ["setuptools>=61.0", "wheel", "numpy==1.26.4"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "picht"
-version = "1.1.3"
+version = "1.1.4"
 description = "Electron and ion optics simulation using the Finite Difference Method (FDM)"
 readme = "README.md"
 requires-python = ">=3.6"
