+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy
+numpy==1.26.4
 matplotlib
 scipy
 numba
