+++ docs/index.rst
@@ -1,0 +1,196 @@
+Picht
+=====
+
+An electron optics library that uses the finite difference method (FDM) to simulate electron and ion trajectories through electrostatic lenses. Currently supports Dirichlet boundary conditions, relativistic energies, non-paraxial beam configurations, and provides pre-packaged support for cylindrical and unipotential (einzel) lenses, as well as scripting for custom lens geometries.
+
+It exists to provide a tool that's free and open-source, easily modifiable, and just as powerful as commercial tools, but with architectural decisions that enable even greater power and accuracy, through intelligent architectural decisions and a focus constrained to electron optics — a branch of computational physics with a relatively small open-source community. Capable of simulating electrostatic lens arrays, electron microscopes (SEM, TEM, STEM), focused ion beams (Ga/He/Ne), and mass spectrometers, and any and all applications, from the classical to relativistic regime that benefit from axisymmetric electron optics simulations without magnetic field handling.
+
+Installation
+------------
+
+To install Picht via pip:
+
+.. code-block:: bash
+
+   pip install picht
+
+.. image:: https://img.shields.io/pypi/v/picht.svg 
+   :target: https://pypi.org/project/picht/ 
+.. image:: https://github.com/rolypolytoy/picht/actions/workflows/tests.yml/badge.svg 
+   :target: https://github.com/rolypolytoy/picht/actions/workflows/tests.yml 
+.. image:: https://img.shields.io/badge/License-MIT-yellow.svg 
+   :target: https://opensource.org/licenses/MIT 
+
+Basic Usage
+-----------
+
+You can initialize an einzel lens as follows:
+
+.. code-block:: python
+
+   import numpy as np
+   from picht import IonOpticsSystem, ElectrodeConfig
+   import matplotlib.pyplot as plt
+
+   system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1) # all grid units are in mm.
+
+   system.add_einzel_lens(
+       position=20.0,
+       width=60.0,
+       aperture_center=50.0,
+       aperture_width=48.0,
+       outer_diameter=50.0,
+       focus_voltage=-7000,
+       gap_size = 4
+   )
+   potential = system.solve_fields()
+
+   trajectories = system.simulate_beam(
+       energy_eV= 10000,  
+       start_z=0,
+       r_range=(0.0499925, 0.0500075),
+       angle_range=(0, 0),
+       num_particles=6,
+       simulation_time=2e-9
+   )
+
+   figure = system.visualize_system(
+       trajectories=trajectories
+   )
+
+   plt.show()
+
+This produces, in less than 30 seconds, a physically realistic map of the particle trajectories.
+
+.. image:: https://github.com/user-attachments/assets/d5f92b58-d0d4-4d68-8d23-6b07bb790105 
+
+In this we can observe several realistic behaviors, including how the fringing fields in einzel lenses first mildly defocus and then focus the beam, the beam crossover, and spherical aberration in the lens.
+
+Advanced Use Case: SEM Simulation
+--------------------------------
+
+Here's a full simulation of an electrostatic lens-only scanning electron microscope (SEM):
+
+.. code-block:: python
+
+   import numpy as np
+   from picht import IonOpticsSystem, ElectrodeConfig
+   import matplotlib.pyplot as plt
+
+   system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1) # all grid units are in mm.
+
+   wehnelt1 = ElectrodeConfig(
+       start=0,
+       width=30,
+       ap_start=30,
[…]
