+++ requirements.txt
@@ -5,3 +5,5 @@ numba
 mendeleev
 pytest
 picht
+sphinx
+sphinx_rtd_theme
