+++ requirements.txt
@@ -5,5 +5,3 @@ numba
 mendeleev
 pytest
 picht
-sphinx
-sphinx_rtd_theme
