### 2025-05-08 #242 @rolypolytoy
Hi @coatless,
I've fixed the issues with v1.1.1 in v1.1.4- I've compiled the release with numpy 1.26.4, and in the dependencies specified the usage of this file. Hopefully this resolves the dependency issues- I've tested it on Google Colab and it seems to be working fine. I did find that after I installed picht on the instance, to restart the session, otherwise it would still act like numpy 2.2.5 was active.

My test document is here:
https://github.com/rolypolytoy/testbed/blob/main/Picht.ipynb

I've also added a user-facing website using Github pages at: https://rolypolytoy.github.io/picht/ using Sphinx. I've kept in the installation instructions and tutorial examples. I've also added much cleaner documentation in the repository, including significantly more comments and docstrings to enable easy modification of the internals by potential contributors or forkers.

Let me know if you find any more bugs or need any more changes to this submission, I'm happy to improve Picht as much as necessary!
