### 2025-05-08 9f2a20c
better documentation and docstrings
