+++ picht/core.py
@@ -17,6 +17,19 @@ class ElectrodeConfig:
 
 @nb.njit
 def solve_field(potential, mask, maxiter, thresh, omega=1.9):
+    """
+    This solves the electrostatic potential using successive over relaxation. This
+    codebase previously used the Jacobi method but we switched to SOR due to greater
+    performance without that much overhead.
+    
+    We implement a numerical solution to Laplace's equation for electrostatic
+    potential in cylindrical coordinates using the SOR technique with Dirichlet boundary
+    conditions.
+
+    I highly recommend not adjusting anything here, including the value for omega, or
+    removing the @nb.njit tag, because this is a performance-critical function- perhaps
+    the most performance-critical one.
+    """
     for _ in range(maxiter):
         vold = potential.copy()
         
@@ -41,6 +54,7 @@ def solve_field(potential, mask, maxiter, thresh, omega=1.9):
             
     return potential
 
+#retrieve electric field components at a given position using interpolation to obtain electric field values at fractional grid positions
 @nb.njit
 def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
     if 0 <= z < axial_size and 0 <= r < radial_size:
@@ -52,6 +66,29 @@ def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
 
 @nb.njit
 def calc_dynamics(z, r, vz, vr, Ez, Er, qm, mass, c):
+    """    
+    Computes relativistic dynamics of charged particles using
+    the Lorentz factor (special relativity) and the Lorentz force 
+    for electrostatics. You can swap the Lorentz force handling
+    with the paraxial ray equation for easy speedups- I've not 
+    done so intentionally to maintain more physical realism.
+    
+    Arguments:
+        z: Z-axis position in meters
+        r: R-axis position in meters
+        vz: Z-axis velocity in meters per second
+        vr: R-axis velocity in meters per second
+        Ez: Z-axis electric field in volts per meter
+        Er: R-axis electric field in volts per meter
+        qm: charge/mass ratio of the particle
+        mass: particle mass- self explanatory
+        c: speed of light in a vacuum
+        
+    Returns:
+        ndarray: Array containing [vz, vr, az, ar], for complete kinematic and kinetic information about the particle.
+
+    Feel free to modify this. What the code does here is intuitive, and if you want to simulate non-relativistic physics simply force gamma = 1.0.
+    """
     vsq = vz**2 + vr**2
     csq = c**2
     
@@ -69,6 +106,24 @@ def calc_dynamics(z, r, vz, vr, Ez, Er, qm, mass, c):
     return np.array([vz, vr, az, ar])
 
 class PotentialField:    
+    """
+    Attributes:
+        nz: Grid points in the z-axis
+        nr: Grid points in the r-axis
+        axial_size: Size of the z-axis in meters. 
+        radial_size: Size of the r-axis in meters. 
+
+        Note: In a lot of places in Picht we use "grid units", where one unit in the z-axis is found by dz = axial_size/nz,
+        and one unit in the r-axis is found by dr = radial_size/nr. Remember this when initializing electrodes or einzel lenses.
+
+        dz: Grid spacing in z-axis
+        dr: Grid spacing in r-axis
+        potential: 2D array with electric potential values
+        electrode_mask: Boolean mask for electrode positions
+        Ez: Z-component of electric field
+        Er: R-component of electric field
+    """
+    
     def __init__(self, nz: float, nr: float, axial_size: float, radial_size: float):
         self.nz = nz
         self.nr = nr
@@ -81,6 +136,7 @@ class PotentialField:
         self.Ez = None
         self.Er = None
     
[…]
