### 2025-05-08 #242 @coatless
@rolypolytoy thanks for the update. I'm still running into version conflicts on Colab. That may be due to some underlying config or test group I'm in.

Running locally, the new versioned packages work:

Regarding the documentation site, there are still a few ways to go. We need the API documentation:

https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/document-your-code-api-docstrings.html

Alongside a package tutorial that differs from the README.md:

https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/create-package-tutorials.html
