### 2025-05-08 026ca39
Update core.py
