+++ picht/core.py
@@ -8,6 +8,17 @@ from mendeleev import element
 
 @dataclass
 class ElectrodeConfig:
+    """
+    Configures the simplest kind of electrostatic lens: the cylindrical electrode.
+    
+    Attributes:
+        start (float): Position where the electrode begins on the z-axis in grid units.
+        width (float): Width of the electrode on the z-axis in grid units.
+        ap_start (float): Starting position of the aperture on the r-axis in grid units.
+        ap_width (float): Width of the aperture on the r-axis in grid units.
+        outer_diameter (float): Full diameter of the electrode on the r-axis in grid units.
+        voltage (float): Voltage of the electrode in volts.
+    """
     start: float
     width: float
     ap_start: float
@@ -18,17 +29,25 @@ class ElectrodeConfig:
 @nb.njit
 def solve_field(potential, mask, maxiter, thresh, omega=1.9):
     """
-    This solves the electrostatic potential using successive over relaxation. This
-    codebase previously used the Jacobi method but we switched to SOR due to greater
-    performance without that much overhead.
-    
-    We implement a numerical solution to Laplace's equation for electrostatic
+    Implements a numerical solution to Laplace's equation for electrostatic
     potential in cylindrical coordinates using the SOR technique with Dirichlet boundary
-    conditions.
-
-    I highly recommend not adjusting anything here, including the value for omega, or
-    removing the @nb.njit tag, because this is a performance-critical function- perhaps
-    the most performance-critical one.
+    conditions. It iteratively refines the potential field until convergence is achieved, as defined
+    by the thresh argument.
+    
+    Parameters:
+        potential (ndarray): 2D numpy array containing initial potential values.
+        mask (ndarray): Boolean mask indicating electrode positions, True where electrodes are.
+        maxiter (int): Maximum number of iterations to perform for the successive over relaxation solver.
+        thresh (float): Convergence threshold for the solution.
+        omega (float, optional): Relaxation parameter for SOR- omega defaults to 1.9 and must be kept between 1 and 2 otherwise SOR is invalid.
+    
+    Returns:
+        ndarray: 2D numpy array containing the solved potential field.
+        
+    Note:
+        This is a performance-critical function optimized with Numba's JIT compilation.
+        Modifying the omega parameter or removing the @nb.njit tag may significantly
+        impact performance.
     """
     for _ in range(maxiter):
         vold = potential.copy()
@@ -54,9 +73,29 @@ def solve_field(potential, mask, maxiter, thresh, omega=1.9):
             
     return potential
 
-#retrieve electric field components at a given position using interpolation to obtain electric field values at fractional grid positions
 @nb.njit
 def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
+    """
+    Provides electric field values at fractional grid positions by
+    interpolating between grid points. It handles boundary conditions by returning
+    zero field values for positions outside the simulation domain, which means
+    outside the defined domain, no forces can be imposed on the charged particles.
+    
+    Parameters:
+        z (float): Z-axis position in meters.
+        r (float): R-axis position in meters.
+        Ez (ndarray): Z-component of the electric field array.
+        Er (ndarray): R-component of the electric field array.
+        axial_size (float): Total size of z-axis in meters.
+        radial_size (float): Total size of r-axis in meters.
+        dz (float): Grid spacing in z-axis, found by dividing nz/z.
+        dr (float): Grid spacing in r-axis, found by dividing nr/r.
+        nz (int): Number of grid points in z-axis.
+        nr (int): Number of grid points in r-axis.
+        
+    Returns:
+        tuple: (Ez, Er) Electric field components at the specified position in V/m.
+    """
[…]
