+++ docs/index.rst
@@ -21,10 +21,16 @@ To install Picht via pip:
 .. image:: https://img.shields.io/badge/License-MIT-yellow.svg 
    :target: https://opensource.org/licenses/MIT 
 
-Basic Usage
+Electrode Creation
 -----------
+Before creating any electrodes, we need to familiarize ourselves with a few concepts. First- whenever you initialize a system, you need to pick 
+nr, nz, axial_size and radial_size. nr is the grid resolution in the r-dimension, and nz is the same for the z-dimension. Axial_size is the total 
+length of the domain in meters in the z-dimension, and radial_size is the same for the r-dimension.
 
-You can initialize an einzel lens as follows:
+This means the grid resolution in the z-dimension is axial_size/nz, and in the r-dimension is radial_size/nr. You should, if using nonstandard values for these,
+ensure you're aware of these so you can properly convert from grid units to meters.
+
+You can initialize a cylindrical electrode as as follows:
 
 .. code-block:: python
 
@@ -32,7 +38,51 @@ You can initialize an einzel lens as follows:
    from picht import IonOpticsSystem, ElectrodeConfig
    import matplotlib.pyplot as plt
 
-   system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1) # all grid units are in mm.
+   system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
+
+   electrode = ElectrodeConfig(
+        start=30,
+        width=5,
+        ap_start=40,
+        ap_width=20,
+        outer_diameter = 50,
+        voltage=-5000
+        )
+   potential = system.solve_fields()
+
+   trajectories = system.simulate_beam(
+       energy_eV= 10000,  
+       start_z=0,
+       r_range=(0.0499925, 0.0500075),
+       angle_range=(0, 0),
+       num_particles=6,
+       simulation_time=2e-9
+   )
+
+   figure = system.visualize_system(
+       trajectories=trajectories
+   )
+
+   plt.show()
+
+Note the syntax for configuring an electrode, and how we parameterize the beam's energy in eV, its start_z, and its radial range. Keep angle_range = (0, 0) for parallel beams, or (-θ, θ), where θ is your intended half-angle of divergence.
+As we can see- this produces some degree of focusing behavior, but after the electrode ends, the beam divergences once more, showcasing why we don't often use cylindrical-only lens elements, despite their simplicity.
+
+Now, we'll build an Einzel Lens at -5000V with a beam energy of 10keV electrons. 
+
+Unipotential Lenses
+--------------------------------
+Einzel lenses are three cylindrical lenses arranged in a specific pattern: the first electrode is grounded (0V), the second is at its focus_voltage, and the third is also grounded at (0V). 
+They're called unipotential lenses because they can provide beam focusing and defocusing without affecting the beam's net energy, which reduces (sometimes!) the behavior we saw in cylindrical electrodes.
+Always make sure the aperture_center (in grid units) aligns with the center of the r_range() when you parameterize your beam.
+
+.. code-block:: python
+
+   import numpy as np
+   from picht import IonOpticsSystem, ElectrodeConfig
+   import matplotlib.pyplot as plt
+
+   system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
 
    system.add_einzel_lens(
        position=20.0,
@@ -40,8 +90,7 @@ You can initialize an einzel lens as follows:
        aperture_center=50.0,
        aperture_width=48.0,
        outer_diameter=50.0,
-       focus_voltage=-7000,
-       gap_size = 4
+       focus_voltage=-5000
    )
[…]
