+++ examples/File.md
@@ -1,1 +1,8 @@
+File dump:
+![electrode](https://github.com/user-attachments/assets/493d152d-b57e-425d-bcfd-545ca7eefac4)
 
+![Focusing](https://github.com/user-attachments/assets/c14983e2-4fa1-4191-93e1-01c6a47082c2)
+
+![Defocusing](https://github.com/user-attachments/assets/2f5890ed-85ba-47f0-988e-8695df49adb8)
+
+![reflection](https://github.com/user-attachments/assets/7d54258c-4a83-4e33-b315-efc9c0d04d37)
