### 2025-05-08 #242 @rolypolytoy
Hi @coatless,
Glad to see the package is working fine. The thinc message came up on my Jupyter notebook too, but it didn't interfere with my package's functioning, and you can safely run it without issues.

As for docstrings- I've added them to my code (https://github.com/rolypolytoy/picht/blob/main/picht/core.py) with better type hints and explanations of what everything does.

I've also added 3 new tutorial examples- to better enable people to create complex lens systems out of the box and avoid common design pitfalls, and I've only added it to the documentation website to provide a proper separation between the README.md and the documentation website, which is more in-depth: https://rolypolytoy.github.io/picht/

Let me know if there are any more changes that need to be made!
