### 2025-05-09 #242 @coatless
@rolypolytoy need docs site to include the API listing and new tutorials.

From our docs, you can see an example of the docstrings being converted to a web API:

https://www.fatiando.org/verde/latest/api/index.html

This requires using an extension:

https://www.pyopensci.org/python-package-guide/documentation/hosting-tools/sphinx-python-package-documentation-tools.html#sphinx-sites-can-be-customized-using-extensions-and-themes

https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html

Lastly, for the new tutorials, you will need to enable Sphinx gallery likely:

https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/create-package-tutorials.html#nbsphinx-tutorials-using-jupyter-notebooks
