+++ docs/conf.py
@@ -1,25 +1,38 @@
 import os
 import sys
+import matplotlib
 
-sys.path.insert(0, os.path.abspath('.'))
+matplotlib.use('agg')
 
-try:
-    import picht
-    release = picht.__version__
-except ImportError:
-    release = '1.1.4'
+sys.path.insert(0, os.path.abspath('..'))
+
+release = '1.1.4'
+version = '.'.join(release.split('.')[:2])
 
 project = 'picht'
 copyright = '2025, Rishiit Sharma'
 author = 'Rishiit Sharma'
-version = '.'.join(release.split('.')[:2])
 
 extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.viewcode',
     'sphinx.ext.napoleon',
     'sphinx.ext.autosummary',
+    'sphinx_gallery.gen_gallery',
 ]
+
+napoleon_google_docstring = True
+napoleon_numpy_docstring = False
+napoleon_include_init_with_doc = True
+napoleon_include_private_with_doc = False
+napoleon_include_special_with_doc = True
+napoleon_use_admonition_for_examples = True
+napoleon_use_admonition_for_notes = True
+napoleon_use_admonition_for_references = False
+napoleon_use_ivar = False
+napoleon_use_param = True
+napoleon_use_rtype = True
+napoleon_attr_annotations = True
 
 templates_path = ['_templates']
 source_suffix = '.rst'
@@ -34,9 +47,31 @@ autodoc_default_options = {
     'members': True,
     'undoc-members': True,
     'private-members': False,
-    'special-members': False,
+    'special-members': '__init__',
     'inherited-members': False,
     'show-inheritance': True,
 }
 
 autosummary_generate = True
+
+sphinx_gallery_conf = {
+    'examples_dirs': '../examples',
+    'gallery_dirs': 'auto_examples',
+    'filename_pattern': 'example_',
+    'plot_gallery': True,
+    'download_all_examples': True,
+    'line_numbers': True,
+    'remove_config_comments': True,
+    'thumbnail_size': (400, 300),
+    'min_reported_time': 0
+}
+
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/3', None),
+    'numpy': ('https://numpy.org/doc/stable/', None),
+    'scipy': ('https://docs.scipy.org/doc/scipy/', None),
+    'matplotlib': ('https://matplotlib.org/stable/', None),
+}
+
+import logging
+logging.getLogger('sphinx_gallery').setLevel(logging.DEBUG)
