+++ examples/README.rst
@@ -1,0 +1,7 @@
+Picht Tutorial
+=====================
+
+This tutorial contains examples demonstrating how to use Picht, moving up from basic cylindrical electrodes, to 
+understanding the properties of unipotential lenses, to finally being able to design and chain complex multi-lens systems.
+
+Begin at electrode creation, work your way through the unipotential lens series, and then try your hand at the SEM example.
