+++ docs/_sources/api.rst.txt
@@ -1,0 +1,8 @@
+API Documentation
+=================
+
+.. automodule:: picht.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :special-members: __init__

[…]
