### 2025-05-09 #242 @rolypolytoy
@coatless
Ive added API documentation using sphinx autodoc at https://rolypolytoy.github.io/picht/api.html, and added sphinx gallery tutorials (5 total) at https://rolypolytoy.github.io/picht/auto_examples/index.html#.
