### 2025-05-09 d605af9
massive improvements in SEM example
