+++ docs/_sources/auto_examples/example_einzellens_focusing.rst.txt
@@ -78,7 +78,7 @@ either negative to negative, or positive to positive. We'll thus set it to -5000
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 4.282 seconds)
+   **Total running time of the script:** (0 minutes 3.017 seconds)
 
 
 .. _sphx_glr_download_auto_examples_example_einzellens_focusing.py:
