+++ README.md
@@ -50,99 +50,10 @@ It produces, in less than 30 seconds, a physically realistic map of the particle
 
 In this we can observe several realistic behaviors, including how the fringing fields in einzel lenses first mildly defocus and then focus the beam, the beam crossover, and spherical aberration in the lens. By default, we assume Dirichlet boundary conditions, to better simulate real electrostatic lens systems with metal boundaries. Neumann boundary conditions might provide more idealized behavior, and are the defaults in most commercial electron optics solvers (including those used in COMSOL and ANSYS Maxwell), however since Dirichlet boundary conditions effectively simulate grounded boundaries rather than infinitely extending ones, for real-life systems, this is significantly more accurate, and reduces the insidious simulation-experimental gap.
 
-Here, we demonstrate a complete simulation of an electrostatic lens-only scanning electron microscope with full accounting of divergence post-acceleration, Wehnelt cylinders, and one condenser lens and objective lens, with the final focal length ~8.7 mm after the final lens- an entirely physically plausible number, with tight convergence. I've increased the amount of particles from 6 to 100, and increased the initial beam divergence from 0 radians to +-2 radians to more accurately model the physical 'boiloff' process of thermionic sources. Regardless, the initial beam is quite straight due to acceleration between the cathode and anode, and we get this in just a few minutes:
+A more complex example looks like this:
+![SEM](https://github.com/user-attachments/assets/6fcb361e-1bca-42dc-8548-b18654829814)
 
-```python
-import numpy as np
-from picht import IonOpticsSystem, ElectrodeConfig
-import matplotlib.pyplot as plt
-
-system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1) #all grid units are in mm.
-
-
-wehnelt1 = ElectrodeConfig(
-    start=0,
-    width=30,
-    ap_start=30,
-    ap_width=40,
-    outer_diameter = 50,
-    voltage=-5100
-)
-wehnelt2 = ElectrodeConfig(
-    start=30,
-    width=5,
-    ap_start=40,
-    ap_width=20,
-    outer_diameter = 50,
-    voltage=-5100
-)
-system.add_electrode(wehnelt1)
-system.add_electrode(wehnelt2)
-anode = ElectrodeConfig(
-    start=40,
-    width = 2,
-    ap_start=48,
-    ap_width=4,
-    outer_diameter = 50,
-    voltage=0
-)
-cathode = ElectrodeConfig(
-    start=22,
-    width = 2,
-    ap_start=50,
-    ap_width=0,
-    outer_diameter = 2,
-    voltage=-5000
-)
-
-system.add_electrode(anode)
-system.add_einzel_lens(
-    position=80.0,
-    width=60.0,
-    aperture_center=50.0,
-    aperture_width=48.0,
-    outer_diameter=50.0,
-    focus_voltage=-7000
-)
-
-system.add_einzel_lens(
-    position=160.0,
-    width=60.0,
-    aperture_center=50.0,
-    aperture_width=48.0,
-    outer_diameter=50.0,
-    focus_voltage=-6500
-)
-potential = system.solve_fields()
-
-trajectories = system.simulate_beam(
-    energy_eV= 10,  
-    start_z=0.025,
-    r_range=(0.0499925, 0.0500075),
-    angle_range=(-2, 2),
-    num_particles=6,
-    simulation_time=1e-8
-)
-
-figure = system.visualize_system(
-    trajectories=trajectories)
-
-plt.show()
-```
-
-![SEM](https://github.com/user-attachments/assets/8e4bc3db-832a-4892-869d-d16839526ebe)
-
[…]
