### 2025-05-13 #242 @coatless
@rolypolytoy It would be good to streamline the navigation bar on the left hand side with a "Getting Started" Section first instead of the API.

https://www.pyopensci.org/python-package-guide/documentation/write-user-documentation/get-started.html#four-elements-of-a-good-open-source-documentation-landing-page

You can see the sidebar example structure here:

https://www.fatiando.org/verde/latest/index.html

The reason for these asks is both users and the reviewers will need a solid starting place to understand your package.
