### 2025-05-13 #242 @rolypolytoy
Of course, I'll make sure the website's a lot more intuitively structured and make the function and usability of Picht a lot more clear, I'll get back to  you with the changes soon.
