+++ picht/core.py
@@ -5,11 +5,22 @@ from dataclasses import dataclass
 from typing import Dict, List, Tuple, Optional, Union, Callable
 import numba as nb
 from mendeleev import element
+<<<<<<< Updated upstream
+=======
+import pyamg
+from joblib import Parallel, delayed
+import multiprocessing
+import os
+>>>>>>> Stashed changes
 
 @dataclass
 class ElectrodeConfig:
     """
+<<<<<<< Updated upstream
     Configures the simplest kind of electrostatic lens: the cylindrical electrode.
+=======
+    Configures a simple cylindrically symmetric electrostatic lens.
+>>>>>>> Stashed changes
     
     Attributes:
         start (float): Position where the electrode begins on the z-axis in grid units.
@@ -77,9 +88,14 @@ def solve_field(potential, mask, maxiter, thresh, omega=1.9):
 def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
     """
     Provides electric field values at fractional grid positions by
+<<<<<<< Updated upstream
     interpolating between grid points. It handles boundary conditions by returning
     zero field values for positions outside the simulation domain, which means
     outside the defined domain, no forces can be imposed on the charged particles.
+=======
+    picking the value at the nearest neighbor. Necessary because 
+    FDM is discretizing by nature.
+>>>>>>> Stashed changes
     
     Parameters:
         z (float): Z-axis position in meters.
@@ -103,26 +119,44 @@ def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
     else:
         return 0.0, 0.0
 
+
 @nb.njit
+<<<<<<< Updated upstream
 def calc_dynamics(z, r, vz, vr, Ez, Er, qm, mass, c):
     """    
     Calculates the acceleration of charged particles by applying
     the Lorentz force with special-relativistic corrections. It calculates the Lorentz factor
     through the variable gamma, and provides correct velocities and accelerations with
     minimal overhead.
+=======
+def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
+    """    
+    Calculates the acceleration of charged particles by applying
+    the Lorentz force with special-relativistic corrections. Uses energy
+    momentum formalism for full eV to TeV support.
+>>>>>>> Stashed changes
     
     Parameters:
         z (float): Z-axis position in meters.
         r (float): R-axis position in meters.
+<<<<<<< Updated upstream
         vz (float): Z-axis velocity in meters per second.
         vr (float): R-axis velocity in meters per second.
         Ez (float): Z-axis electric field in volts per meter.
         Er (float): R-axis electric field in volts per meter.
         qm (float): Charge-to-mass ratio of the particle.
+=======
+        pz (float): Z-axis momentum in kg * m/s.
+        pr (float): R-axis momentum in kg * m/s.
+        Ez (float): Z-axis electric field in volts per meter.
+        Er (float): R-axis electric field in volts per meter.
+        q (float): Charge of the particle.
+>>>>>>> Stashed changes
         mass (float): Particle mass in kg.
         c (float): Speed of light in a vacuum in meters per second.
         
     Returns:
+<<<<<<< Updated upstream
         ndarray: Array containing [vz, vr, az, ar], representing velocity and acceleration
                 components in the z and r directions for complete kinematic information.
     """
@@ -148,6 +182,23 @@ class PotentialField:
     and solving for the electric field. Calculating the electric field is necessary at the start
     of the particle trajectory calculations, as this is the bulk of the information particles use
     for trajectory calculations.
+=======
+        ndarray: Array containing [vz, vr, dpz_dt, dpr_dt], representing velocity and force
+                components in the z and r directions for complete kinematic information (force is dp/dt).
+    """
[…]
