+++ docs/conf.py
@@ -1,25 +1,21 @@
 import os
 import sys
-import matplotlib
-
-matplotlib.use('agg')
-
-sys.path.insert(0, os.path.abspath('..'))
+import matplotlib 
+matplotlib.use('agg') 
 
 release = '1.1.4'
-version = '.'.join(release.split('.')[:2])
+version = '.'.join(release.split('.')[:2]) 
 
 project = 'picht'
 copyright = '2025, Rishiit Sharma'
-author = 'Rishiit Sharma'
+author = 'Rishiit Sharma' 
 
 extensions = [
-    'sphinx.ext.autodoc',
     'sphinx.ext.viewcode',
     'sphinx.ext.napoleon',
     'sphinx.ext.autosummary',
     'sphinx_gallery.gen_gallery',
-]
+] 
 
 napoleon_google_docstring = True
 napoleon_numpy_docstring = False
@@ -32,16 +28,16 @@ napoleon_use_admonition_for_references = False
 napoleon_use_ivar = False
 napoleon_use_param = True
 napoleon_use_rtype = True
-napoleon_attr_annotations = True
+napoleon_attr_annotations = True 
 
 templates_path = ['_templates']
 source_suffix = '.rst'
 master_doc = 'index'
 language = 'en'
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store'] 
 
 html_theme = 'sphinx_rtd_theme'
-html_static_path = ['_static']
+html_static_path = ['_static'] 
 
 autodoc_default_options = {
     'members': True,
@@ -50,9 +46,9 @@ autodoc_default_options = {
     'special-members': '__init__',
     'inherited-members': False,
     'show-inheritance': True,
-}
+} 
 
-autosummary_generate = True
+autosummary_generate = True 
 
 sphinx_gallery_conf = {
     'examples_dirs': '../examples',
@@ -64,14 +60,14 @@ sphinx_gallery_conf = {
     'remove_config_comments': True,
     'thumbnail_size': (400, 300),
     'min_reported_time': 0
-}
+} 
 
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3', None),
     'numpy': ('https://numpy.org/doc/stable/', None),
     'scipy': ('https://docs.scipy.org/doc/scipy/', None),
     'matplotlib': ('https://matplotlib.org/stable/', None),
-}
+} 
 
 import logging
 logging.getLogger('sphinx_gallery').setLevel(logging.DEBUG)
