+++ requirements.txt
@@ -1,7 +1,13 @@
-numpy==1.26.4
 matplotlib
 scipy
 numba
 mendeleev
 pytest
+<<<<<<< Updated upstream
 picht
+=======
+pyamg
+joblib
+picht==1.1.4
+numpy==1.26.4
+>>>>>>> Stashed changes
