+++ docs/_sources/auto_examples/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:17.213** total execution time for 5 files **from auto_examples**:
+**00:18.010** total execution time for 5 files **from auto_examples**:
 
 .. container::
 
@@ -33,17 +33,17 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_auto_examples_example_sem_simulation.py` (``example_sem_simulation.py``)
-     - 00:06.260
+     - 00:05.972
      - 0.0
    * - :ref:`sphx_glr_auto_examples_example_einzellens_defocusing.py` (``example_einzellens_defocusing.py``)
-     - 00:03.024
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_example_einzellens_focusing.py` (``example_einzellens_focusing.py``)
-     - 00:03.017
+     - 00:03.090
      - 0.0
    * - :ref:`sphx_glr_auto_examples_example_cylindrical_lens.py` (``example_cylindrical_lens.py``)
-     - 00:02.985
+     - 00:03.086
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_example_einzellens_focusing.py` (``example_einzellens_focusing.py``)
+     - 00:02.995
      - 0.0
    * - :ref:`sphx_glr_auto_examples_example_einzellens_deflection.py` (``example_einzellens_deflection.py``)
-     - 00:01.928
+     - 00:02.868
      - 0.0

[…]
