+++ picht/core.py
@@ -11,13 +11,20 @@ import pyamg
 from joblib import Parallel, delayed
 import multiprocessing
 import os
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 
 @dataclass
 class ElectrodeConfig:
     """
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
     Configures the simplest kind of electrostatic lens: the cylindrical electrode.
+=======
+    Configures a simple cylindrically symmetric electrostatic lens.
+>>>>>>> Stashed changes
 =======
     Configures a simple cylindrically symmetric electrostatic lens.
 >>>>>>> Stashed changes
@@ -89,9 +96,14 @@ def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
     """
     Provides electric field values at fractional grid positions by
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
     interpolating between grid points. It handles boundary conditions by returning
     zero field values for positions outside the simulation domain, which means
     outside the defined domain, no forces can be imposed on the charged particles.
+=======
+    picking the value at the nearest neighbor. Necessary because 
+    FDM is discretizing by nature.
+>>>>>>> Stashed changes
 =======
     picking the value at the nearest neighbor. Necessary because 
     FDM is discretizing by nature.
@@ -134,11 +146,15 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
     Calculates the acceleration of charged particles by applying
     the Lorentz force with special-relativistic corrections. Uses energy
     momentum formalism for full eV to TeV support.
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
     
     Parameters:
         z (float): Z-axis position in meters.
         r (float): R-axis position in meters.
+<<<<<<< Updated upstream
 <<<<<<< Updated upstream
         vz (float): Z-axis velocity in meters per second.
         vr (float): R-axis velocity in meters per second.
@@ -146,16 +162,22 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
         Er (float): R-axis electric field in volts per meter.
         qm (float): Charge-to-mass ratio of the particle.
 =======
+=======
+>>>>>>> Stashed changes
         pz (float): Z-axis momentum in kg * m/s.
         pr (float): R-axis momentum in kg * m/s.
         Ez (float): Z-axis electric field in volts per meter.
         Er (float): R-axis electric field in volts per meter.
         q (float): Charge of the particle.
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
         mass (float): Particle mass in kg.
         c (float): Speed of light in a vacuum in meters per second.
         
     Returns:
+<<<<<<< Updated upstream
 <<<<<<< Updated upstream
         ndarray: Array containing [vz, vr, az, ar], representing velocity and acceleration
                 components in the z and r directions for complete kinematic information.
@@ -183,6 +205,8 @@ class PotentialField:
     of the particle trajectory calculations, as this is the bulk of the information particles use
     for trajectory calculations.
 =======
+=======
+>>>>>>> Stashed changes
         ndarray: Array containing [vz, vr, dpz_dt, dpr_dt], representing velocity and force
                 components in the z and r directions for complete kinematic information (force is dp/dt).
     """
@@ -198,6 +222,9 @@ class PotentialField:
     """
     This class handles everything related to the electric potential field, including
     adding electrodes and einzel lenses, and solving for the electric field.
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
     
[…]
