+++ requirements.txt
@@ -10,4 +10,7 @@ pyamg
 joblib
 picht==1.1.4
 numpy==1.26.4
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes

[…]
