+++ picht/core.py
@@ -12,6 +12,9 @@ from joblib import Parallel, delayed
 import multiprocessing
 import os
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
@@ -21,7 +24,11 @@ class ElectrodeConfig:
     """
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
     Configures the simplest kind of electrostatic lens: the cylindrical electrode.
+=======
+    Configures a simple cylindrically symmetric electrostatic lens.
+>>>>>>> Stashed changes
 =======
     Configures a simple cylindrically symmetric electrostatic lens.
 >>>>>>> Stashed changes
@@ -97,9 +104,14 @@ def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
     Provides electric field values at fractional grid positions by
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
     interpolating between grid points. It handles boundary conditions by returning
     zero field values for positions outside the simulation domain, which means
     outside the defined domain, no forces can be imposed on the charged particles.
+=======
+    picking the value at the nearest neighbor. Necessary because 
+    FDM is discretizing by nature.
+>>>>>>> Stashed changes
 =======
     picking the value at the nearest neighbor. Necessary because 
     FDM is discretizing by nature.
@@ -147,6 +159,9 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
     the Lorentz force with special-relativistic corrections. Uses energy
     momentum formalism for full eV to TeV support.
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
@@ -154,6 +169,7 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
     Parameters:
         z (float): Z-axis position in meters.
         r (float): R-axis position in meters.
+<<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
         vz (float): Z-axis velocity in meters per second.
@@ -164,12 +180,17 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
 =======
 =======
 >>>>>>> Stashed changes
+=======
+>>>>>>> Stashed changes
         pz (float): Z-axis momentum in kg * m/s.
         pr (float): R-axis momentum in kg * m/s.
         Ez (float): Z-axis electric field in volts per meter.
         Er (float): R-axis electric field in volts per meter.
         q (float): Charge of the particle.
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
@@ -177,6 +198,7 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
         c (float): Speed of light in a vacuum in meters per second.
         
     Returns:
+<<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
         ndarray: Array containing [vz, vr, az, ar], representing velocity and acceleration
@@ -207,6 +229,8 @@ class PotentialField:
 =======
 =======
 >>>>>>> Stashed changes
+=======
+>>>>>>> Stashed changes
         ndarray: Array containing [vz, vr, dpz_dt, dpr_dt], representing velocity and force
[…]
