+++ requirements.txt
@@ -11,6 +11,9 @@ joblib
 picht==1.1.4
 numpy==1.26.4
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
