+++ pyproject.toml
@@ -6,7 +6,11 @@ build-backend = "setuptools.build_meta"
 name = "picht"
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
 version = "1.1.4"
+=======
+version = "2.0.0"
+>>>>>>> Stashed changes
 =======
 version = "2.0.0"
 >>>>>>> Stashed changes
@@ -35,6 +39,9 @@ dependencies = [
     "h5py",
     "numpy==1.26.4",
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
