+++ picht/core.py
@@ -13,6 +13,9 @@ import multiprocessing
 import os
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
@@ -25,7 +28,11 @@ class ElectrodeConfig:
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
     Configures the simplest kind of electrostatic lens: the cylindrical electrode.
+=======
+    Configures a simple cylindrically symmetric electrostatic lens.
+>>>>>>> Stashed changes
 =======
     Configures a simple cylindrically symmetric electrostatic lens.
 >>>>>>> Stashed changes
@@ -105,9 +112,14 @@ def get_field(z, r, Ez, Er, axial_size, radial_size, dz, dr, nz, nr):
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
     interpolating between grid points. It handles boundary conditions by returning
     zero field values for positions outside the simulation domain, which means
     outside the defined domain, no forces can be imposed on the charged particles.
+=======
+    picking the value at the nearest neighbor. Necessary because 
+    FDM is discretizing by nature.
+>>>>>>> Stashed changes
 =======
     picking the value at the nearest neighbor. Necessary because 
     FDM is discretizing by nature.
@@ -160,6 +172,9 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
     momentum formalism for full eV to TeV support.
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
@@ -169,6 +184,7 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
     Parameters:
         z (float): Z-axis position in meters.
         r (float): R-axis position in meters.
+<<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
@@ -182,6 +198,8 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
+=======
+>>>>>>> Stashed changes
         pz (float): Z-axis momentum in kg * m/s.
         pr (float): R-axis momentum in kg * m/s.
         Ez (float): Z-axis electric field in volts per meter.
@@ -189,6 +207,9 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
         q (float): Charge of the particle.
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
@@ -198,6 +219,7 @@ def calc_dynamics(z, r, pz, pr, Ez, Er, q, m, c):
         c (float): Speed of light in a vacuum in meters per second.
         
     Returns:
+<<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
@@ -231,6 +253,8 @@ class PotentialField:
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
+=======
+>>>>>>> Stashed changes
[…]
