+++ requirements.txt
@@ -12,6 +12,9 @@ picht==1.1.4
 numpy==1.26.4
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
