+++ pyproject.toml
@@ -7,7 +7,11 @@ name = "picht"
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
 version = "1.1.4"
+=======
+version = "2.0.0"
+>>>>>>> Stashed changes
 =======
 version = "2.0.0"
 >>>>>>> Stashed changes
@@ -40,6 +44,9 @@ dependencies = [
     "numpy==1.26.4",
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
