+++ pyproject.toml
@@ -8,7 +8,11 @@ name = "picht"
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
 version = "1.1.4"
+=======
+version = "2.0.0"
+>>>>>>> Stashed changes
 =======
 version = "2.0.0"
 >>>>>>> Stashed changes
@@ -45,6 +49,9 @@ dependencies = [
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
 <<<<<<< Updated upstream
+<<<<<<< Updated upstream
+>>>>>>> Stashed changes
+=======
 >>>>>>> Stashed changes
 =======
 >>>>>>> Stashed changes
