+++ picht/core.py
@@ -1,54 +1,22 @@
 import numpy as np
 import matplotlib.pyplot as plt
+from matplotlib.animation import FuncAnimation
+from matplotlib.widgets import CheckButtons, Slider
 from scipy.integrate import solve_ivp
+from scipy.sparse import diags, csr_matrix, lil_matrix
 from dataclasses import dataclass
 from typing import Dict, List, Tuple, Optional, Union, Callable
 import numba as nb
 from mendeleev import element
-<<<<<<< Updated upstream
-=======
 import pyamg
 from joblib import Parallel, delayed
 import multiprocessing
 import os
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
 
 @dataclass
 class ElectrodeConfig:
     """
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-    Configures the simplest kind of electrostatic lens: the cylindrical electrode.
-=======
     Configures a simple cylindrically symmetric electrostatic lens.
->>>>>>> Stashed changes
-=======
-    Configures a simple cylindrically symmetric electrostatic lens.
->>>>>>> Stashed changes
-=======
-    Configures a simple cylindrically symmetric electrostatic lens.
->>>>>>> Stashed changes
-=======
-    Configures a simple cylindrically symmetric electrostatic lens.
->>>>>>> Stashed changes
-=======
-    Configures a simple cylindrically symmetric electrostatic lens.
->>>>>>> Stashed changes
     
     Attributes:
         start (float): Position where the electrode begins on the z-axis in grid units.
@@ -66,84 +34,11 @@ class ElectrodeConfig:
     voltage: float
 
 @nb.njit
-def solve_field(potential, mask, maxiter, thresh, omega=1.9):
-    """
-    Implements a numerical solution to Laplace's equation for electrostatic
-    potential in cylindrical coordinates using the SOR technique with Dirichlet boundary
-    conditions. It iteratively refines the potential field until convergence is achieved, as defined
-    by the thresh argument.
-    
-    Parameters:
-        potential (ndarray): 2D numpy array containing initial potential values.
-        mask (ndarray): Boolean mask indicating electrode positions, True where electrodes are.
-        maxiter (int): Maximum number of iterations to perform for the successive over relaxation solver.
-        thresh (float): Convergence threshold for the solution.
-        omega (float, optional): Relaxation parameter for SOR- omega defaults to 1.9 and must be kept between 1 and 2 otherwise SOR is invalid.
-    
-    Returns:
-        ndarray: 2D numpy array containing the solved potential field.
-        
-    Note:
-        This is a performance-critical function optimized with Numba's JIT compilation.
-        Modifying the omega parameter or removing the @nb.njit tag may significantly
-        impact performance.
-    """
-    for _ in range(maxiter):
-        vold = potential.copy()
-        
-        for i in range(1, potential.shape[0]-1):
-            for j in range(1, potential.shape[1]-1):
-                if not mask[i, j]:
-                    v_new = 0.25 * (
-                        potential[i+1, j] + potential[i-1, j] +
-                        potential[i, j+1] + potential[i, j-1]
-                    )
-                    potential[i, j] = vold[i, j] + omega * (v_new - vold[i, j])
-        
-        maxdiff = 0.0
-        for i in range(potential.shape[0]):
[…]
