+++ picht/example.py
@@ -1,195 +1,0 @@
-<<<<<<< Updated upstream
-=======
-"""
-Unipotential Lenses: Focusing
---------------------------------
-Einzel lenses are three cylindrical lenses arranged in a specific pattern: the first electrode is grounded (0V), the second is at its focus_voltage, and the third is also grounded at (0V). 
-They're called unipotential lenses because they can provide beam focusing and defocusing without affecting the beam's net energy, which reduces (sometimes!) the behavior we saw in cylindrical electrodes.
-Always make sure the aperture_center (in grid units) aligns with the center of the r_range() when you parameterize your beam.
-
-We'll first examine an example of an einzel lens used for focusing, which happens when the polarity of the focus_voltage is the same as the polarity of the charge-
-either negative to negative, or positive to positive. We'll thus set it to -5000V, with initial electron energies at 10keV.
-"""
-
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
-=======
->>>>>>> Stashed changes
-import numpy as np
-from core import IonOpticsSystem, ElectrodeConfig
-import matplotlib.pyplot as plt
-
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-<<<<<<< Updated upstream
-system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1) #all grid units are in mm.
-
-
-wehnelt1 = ElectrodeConfig(
-    start=0,
-    width=30,
-    ap_start=30,
-    ap_width=40,
-    outer_diameter = 50,
-    voltage=-5100
-)
-wehnelt2 = ElectrodeConfig(
-    start=30,
-    width=5,
-    ap_start=40,
-    ap_width=20,
-    outer_diameter = 50,
-    voltage=-5100
-)
-system.add_electrode(wehnelt1)
-system.add_electrode(wehnelt2)
-anode = ElectrodeConfig(
-    start=40,
-    width = 2,
-    ap_start=48,
-    ap_width=4,
-    outer_diameter = 50,
-    voltage=0
-)
-cathode = ElectrodeConfig(
-    start=22,
-    width = 2,
-    ap_start=50,
-    ap_width=0,
-    outer_diameter = 2,
-    voltage=-5000
-)
-
-system.add_electrode(anode)
-system.add_einzel_lens(
-    position=80.0,
-=======
-system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
-
-
-system.add_einzel_lens(
-    position=20.0,
->>>>>>> Stashed changes
-=======
-system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
-
-
-system.add_einzel_lens(
-    position=20.0,
->>>>>>> Stashed changes
-=======
-system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
-
-
-system.add_einzel_lens(
-    position=20.0,
->>>>>>> Stashed changes
-=======
-system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
-
-
-system.add_einzel_lens(
-    position=20.0,
[…]
