+++ requirements.txt
@@ -2,10 +2,6 @@ matplotlib
 scipy
 numba
 mendeleev
-pyamg
-joblib
-cadquery
-h5py
 picht==1.1.4
 pytest
 numpy==1.26.4
