+++ docs/auto_examples/example_cylindrical_lens.rst
@@ -1,114 +1,0 @@
-
-.. DO NOT EDIT.
-.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
-.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
-.. "auto_examples\example_cylindrical_lens.py"
-.. LINE NUMBERS ARE GIVEN BELOW.
-
-.. only:: html
-
-    .. note::
-        :class: sphx-glr-download-link-note
-
-        :ref:`Go to the end <sphx_glr_download_auto_examples_example_cylindrical_lens.py>`
-        to download the full example code.
-
-.. rst-class:: sphx-glr-example-title
-
-.. _sphx_glr_auto_examples_example_cylindrical_lens.py:
-
-
-Electrode Creation
--------------------
-Before creating any electrodes, we need to familiarize ourselves with a few concepts. First- whenever you initialize a system, you need to pick 
-nr, nz, axial_size and radial_size. nr is the grid resolution in the r-dimension, and nz is the same for the z-dimension. Axial_size is the total 
-length of the domain in meters in the z-dimension, and radial_size is the same for the r-dimension.
-
-This means the grid resolution in the z-dimension is axial_size/nz, and in the r-dimension is radial_size/nr. You should, if using nonstandard values for these,
-ensure you're aware of these so you can properly convert from grid units to meters.
-
-You can initialize a cylindrical electrode with -5000V focus voltage and 10keV energy per electron as follows:
-
-.. GENERATED FROM PYTHON SOURCE LINES 13-49
-
-
-
-.. image-sg:: /auto_examples/images/sphx_glr_example_cylindrical_lens_001.png
-   :alt: Electron Trajectories
-   :srcset: /auto_examples/images/sphx_glr_example_cylindrical_lens_001.png
-   :class: sphx-glr-single-img
-
-
-
-
-
-.. code-block:: Python
-   :lineno-start: 13
-
-    import numpy as np
-    from picht import IonOpticsSystem, ElectrodeConfig
-    import matplotlib.pyplot as plt
-
-    #Specifies the grid size, resolution, and domain
-    system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
-
-    #Configures electrode parameterization
-    electrode = ElectrodeConfig(
-        start=30,
-        width=5,
-        ap_start=40,
-        ap_width=20,
-        outer_diameter = 50,
-        voltage=-5000
-    )
-
-    system.add_electrode(electrode)
-
-    #FDM field solving stage
-    potential = system.solve_fields()
-
-    #Beam parameterization
-    trajectories = system.simulate_beam(
-        energy_eV= 10000,  
-        start_z=0,
-        r_range=(0.0499925, 0.0500075),
-        angle_range=(0, 0),
-        num_particles=6,
-        simulation_time=2e-9
-    )
-
-    #Visualization control
-    figure = system.visualize_system(
-        trajectories=trajectories)
-
-    plt.show()
-
-.. rst-class:: sphx-glr-timing
-
-   **Total running time of the script:** (0 minutes 3.063 seconds)
-
-
-.. _sphx_glr_download_auto_examples_example_cylindrical_lens.py:
-
-.. only:: html
-
-  .. container:: sphx-glr-footer sphx-glr-footer-example
-
-    .. container:: sphx-glr-download sphx-glr-download-jupyter
-
-      :download:`Download Jupyter notebook: example_cylindrical_lens.ipynb <example_cylindrical_lens.ipynb>`
-
-    .. container:: sphx-glr-download sphx-glr-download-python
-
[…]
