+++ docs/auto_examples/sg_execution_times.rst
@@ -1,49 +1,0 @@
-
-:orphan:
-
-.. _sphx_glr_auto_examples_sg_execution_times:
-
-
-Computation times
-=================
-**00:17.926** total execution time for 5 files **from auto_examples**:
-
-.. container::
-
-  .. raw:: html
-
-    <style scoped>
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
-    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
-    </style>
-    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
-    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
-    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
-    <script type="text/javascript" class="init">
-    $(document).ready( function () {
-        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
-    } );
-    </script>
-
-  .. list-table::
-   :header-rows: 1
-   :class: table table-striped sg-datatable
-
-   * - Example
-     - Time
-     - Mem (MB)
-   * - :ref:`sphx_glr_auto_examples_example_sem_simulation.py` (``example_sem_simulation.py``)
-     - 00:05.917
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_example_einzellens_defocusing.py` (``example_einzellens_defocusing.py``)
-     - 00:03.098
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_example_cylindrical_lens.py` (``example_cylindrical_lens.py``)
-     - 00:03.063
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_example_einzellens_focusing.py` (``example_einzellens_focusing.py``)
-     - 00:02.966
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_example_einzellens_deflection.py` (``example_einzellens_deflection.py``)
-     - 00:02.882
-     - 0.0
