+++ docs/_sources/auto_examples/example_einzellens_defocusing.rst.txt
@@ -1,103 +1,0 @@
-
-.. DO NOT EDIT.
-.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
-.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
-.. "auto_examples\example_einzellens_defocusing.py"
-.. LINE NUMBERS ARE GIVEN BELOW.
-
-.. only:: html
-
-    .. note::
-        :class: sphx-glr-download-link-note
-
-        :ref:`Go to the end <sphx_glr_download_auto_examples_example_einzellens_defocusing.py>`
-        to download the full example code.
-
-.. rst-class:: sphx-glr-example-title
-
-.. _sphx_glr_auto_examples_example_einzellens_defocusing.py:
-
-
-Unipotential Lenses: Defocusing
---------------------------------
-Here, we can see valid focusing behavior, which happens when the focus voltage is negative, and the particle is negatively charged, or vice versa. 
-We can thus make an einzel lens that causes divergent behavior by re-using the same code but flipping the sign on focus_voltage from -5000 to 5000.
-We can see clearly visible defocusing, demonstrating how einzel lenses are cleanly usable for both focusing and defocusing with single-character changes in the code:
-
-.. GENERATED FROM PYTHON SOURCE LINES 8-38
-
-
-
-.. image-sg:: /auto_examples/images/sphx_glr_example_einzellens_defocusing_001.png
-   :alt: Electron Trajectories
-   :srcset: /auto_examples/images/sphx_glr_example_einzellens_defocusing_001.png
-   :class: sphx-glr-single-img
-
-
-
-
-
-.. code-block:: Python
-   :lineno-start: 8
-
-    import numpy as np
-    from picht import IonOpticsSystem, ElectrodeConfig
-    import matplotlib.pyplot as plt
-
-    system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
-
-
-    system.add_einzel_lens(
-        position=20.0,
-        width=60.0,
-        aperture_center=50.0,
-        aperture_width=48.0,
-        outer_diameter=50.0,
-        focus_voltage=5000
-    )
-
-    potential = system.solve_fields()
-
-    trajectories = system.simulate_beam(
-        energy_eV= 10000,  
-        start_z=0,
-        r_range=(0.0499925, 0.0500075),
-        angle_range=(0, 0),
-        num_particles=6,
-        simulation_time=2e-9
-    )
-
-    figure = system.visualize_system(
-        trajectories=trajectories)
-
-    plt.show()
-
-.. rst-class:: sphx-glr-timing
-
-   **Total running time of the script:** (0 minutes 3.098 seconds)
-
-
-.. _sphx_glr_download_auto_examples_example_einzellens_defocusing.py:
-
-.. only:: html
-
-  .. container:: sphx-glr-footer sphx-glr-footer-example
-
-    .. container:: sphx-glr-download sphx-glr-download-jupyter
-
-      :download:`Download Jupyter notebook: example_einzellens_defocusing.ipynb <example_einzellens_defocusing.ipynb>`
-
-    .. container:: sphx-glr-download sphx-glr-download-python
-
-      :download:`Download Python source code: example_einzellens_defocusing.py <example_einzellens_defocusing.py>`
-
-    .. container:: sphx-glr-download sphx-glr-download-zip
-
-      :download:`Download zipped: example_einzellens_defocusing.zip <example_einzellens_defocusing.zip>`
-
-
-.. only:: html
-
- .. rst-class:: sphx-glr-signature
-
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
