+++ docs/_sources/api.rst.txt
@@ -1,425 +1,0 @@
-API Documentation
-=================
-
-This section provides complete API documentation for all of Picht's internal classes, functions and methods.
-
-Core Classes
-------------
-
-ElectrodeConfig
-~~~~~~~~~~~~~~~
-
-.. class:: ElectrodeConfig
-
-   Dataclass for configuring a cylindrically symmetric electrostatic lens element.
-   
-   **Parameters**
-   
-   :param float start: Z-axis starting position in grid units
-   :param float width: Z-axis width in grid units
-   :param float ap_start: R-axis aperture starting position in grid units
-   :param float ap_width: R-axis aperture width in grid units
-   :param float outer_diameter: Full electrode diameter in grid units
-   :param float voltage: Applied voltage in volts
-
-PotentialField
-~~~~~~~~~~~~~~
-
-.. class:: PotentialField(nz, nr, axial_size, radial_size)
-
-   Manages electric potential fields, electrode configurations, and field solving operations.
-   
-   **Parameters**
-   
-   :param float nz: Number of z-axis grid points
-   :param float nr: Number of r-axis grid points
-   :param float axial_size: Physical z-axis length in meters
-   :param float radial_size: Physical r-axis length in meters
-   
-   **Attributes**
-   
-   .. attribute:: nz
-      :type: int
-      
-      Number of z-axis grid points
-   
-   .. attribute:: nr
-      :type: int
-      
-      Number of r-axis grid points
-   
-   .. attribute:: axial_size
-      :type: float
-      
-      Z-axis size in meters
-   
-   .. attribute:: radial_size
-      :type: float
-      
-      R-axis size in meters
-   
-   .. attribute:: dz
-      :type: float
-      
-      Z-axis grid spacing (meters), calculated as axial_size/nz
-   
-   .. attribute:: dr
-      :type: float
-      
-      R-axis grid spacing (meters), calculated as radial_size/nr
-   
-   .. attribute:: potential
-      :type: numpy.ndarray
-      
-      2D array of electric potential values in volts
-   
-   .. attribute:: electrode_mask
-      :type: numpy.ndarray
-      
-      Boolean mask indicating electrode positions
-   
-   .. attribute:: Ez
-      :type: numpy.ndarray
-      
-      Z-component of electric field in V/m
-   
-   .. attribute:: Er
-      :type: numpy.ndarray
-      
-      R-component of electric field in V/m
-   
-   **Methods**
-   
-   .. method:: add_electrode(config)
-   
-      Adds an electrode to the field configuration.
-      
-      :param ElectrodeConfig config: Electrode configuration parameters
-   
-   .. method:: build_laplacian_matrix(mask, dirichlet_values=None)
-   
-      Constructs sparse matrix representation of the Laplacian operator for solving Laplace's equation.
-      Implements Dirichlet boundary conditions (0V) at all boundaries to simulate grounded metal boundaries.
-      
-      :param numpy.ndarray mask: Boolean array marking electrode positions
-      :param numpy.ndarray dirichlet_values: Optional potential values at masked positions
-      :returns: Sparse matrix A and right-hand side vector b
-      :rtype: tuple(scipy.sparse.csr_matrix, numpy.ndarray)
-   
-   .. method:: solve_potential(max_iterations=500, convergence_threshold=1e-6)
-   
-      Solves the electrostatic potential field using PyAMG multigrid methods.
-      Complexity is O(N) with grid size.
-      
-      :param float max_iterations: Maximum solver iterations (default: 500)
-      :param float convergence_threshold: Convergence criterion (default: 1e-6)
-      :returns: Solved potential field array
-      :rtype: numpy.ndarray
-   
[…]
