+++ docs/_sources/gettingstarted.rst.txt
@@ -1,0 +1,282 @@
+===============
+Getting Started
+===============
+
+**Picht** is a Python library for simulating charged particle optics systems. It provides tools for designing, simulating, and visualizing ion and electron beam trajectories through electrostatic lenses.
+Systems like electron microscopes, focused ion beam systems, and mass spectrometers are difficult to design. A high precision is required to design lens arrays for these systems.
+However, there exist no closed-loop analytical solutions for these systems. To solve this, Picht uses finite difference methods and solves them with the extremely rapid PyAMG multigrid solver. 
+This allows for rapid simulation and prototyping of systems which require careful and precise design of electrostatic lens arrays. 
+
+Installation
+============
+
+Prerequisites
+-------------
+
+Picht requires Python 3.7+ and the following dependencies:
+
+.. code-block:: bash
+
+   numpy
+   matplotlib
+   scipy
+   numba
+   mendeleev
+   pyamg
+   joblib
+   h5py
+   cadquery
+
+You can install Picht and all its dependencies using pip:
+
+.. code-block:: bash
+
+   pip install picht
+
+Quick Start Example
+===================
+
+Here's a simple example that creates an Einzel lens system and simulates electron trajectories:
+
+.. code-block:: python
+
+    import numpy as np
+    from picht import IonOpticsSystem, ElectrodeConfig
+    import matplotlib.pyplot as plt
+
+    system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1)
+
+    system.add_einzel_lens(
+        position=20.0,
+        width=60.0,
+        aperture_center=50.0,
+        aperture_width=48.0,
+        outer_diameter=50.0,
+        focus_voltage=-5000
+    )
+
+    potential = system.solve_fields()
+
+    trajectories = system.simulate_beam(
+        energy_eV= 100,  
+        start_z=0,
+        r_range=(0.0499925, 0.0500075),
+        angle_range=(0, 0),
+        num_particles=6,
+        simulation_time=2e-9
+    )
+
+    figure = system.visualize_system(
+        trajectories=trajectories)
+
+    plt.show()
+
+Basic Concepts
+==============
+
+Grid System
+-----------
+
+Picht uses a cylindrically symmetric grid system:
+
+- **z-axis**: The optical axis, which is the direction the particle beam propagates
+- **r-axis**: The radial distance from the optical axis
+- Grid dimensions are specified in grid units, with physical size determined by ``axial_size`` and ``radial_size`` parameters. They're the system's reference frame and the smallest discrete units where unique potential field values are calculated.
+
+The axisymmetric view, rather than a full 3D treatment, is done to increase computational speeds while maintaining physical accuracy for rotationally symmetric electron optics systems.
+
+Electrode Configuration
+-----------------------
+
+Electrodes are defined using the ``ElectrodeConfig`` dataclass:
+
+.. code-block:: python
+
+   from picht import ElectrodeConfig
+
+   electrode = ElectrodeConfig(
+       start=10,              # Starting position on z-axis (grid units)
+       width=20,              # Width on z-axis (grid units)
+       ap_start=30,           # Aperture start on r-axis (grid units)
+       ap_width=40,           # Aperture width on r-axis (grid units)
+       outer_diameter=100,    # Full diameter (grid units)
+       voltage=500            # Applied voltage (V)
+   )
+
+   system.add_electrode(electrode)
+
+This creates a radially symmetric electrode with an inner diameter equal to ap_width * (radial_size/nr) in meters, and an outer diameter equal to outer_diameter * (radial_size/nr). Using grid units rather than meters feels clunky now, but as you use the system more you'll find it more intuitive.
[…]
