+++ docs/_sources/auto_examples/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:18.877** total execution time for 5 files **from auto_examples**:
+**00:19.038** total execution time for 5 files **from auto_examples**:
 
 .. container::
 
@@ -33,17 +33,17 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_auto_examples_example_sem_simulation.py` (``example_sem_simulation.py``)
-     - 00:06.517
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_example_einzellens_defocusing.py` (``example_einzellens_defocusing.py``)
-     - 00:03.222
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_example_cylindrical_lens.py` (``example_cylindrical_lens.py``)
-     - 00:03.203
+     - 00:06.411
      - 0.0
    * - :ref:`sphx_glr_auto_examples_example_einzellens_focusing.py` (``example_einzellens_focusing.py``)
-     - 00:03.056
+     - 00:03.307
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_example_einzellens_defocusing.py` (``example_einzellens_defocusing.py``)
+     - 00:03.184
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_example_cylindrical_lens.py` (``example_cylindrical_lens.py``)
+     - 00:03.161
      - 0.0
    * - :ref:`sphx_glr_auto_examples_example_einzellens_deflection.py` (``example_einzellens_deflection.py``)
-     - 00:02.879
+     - 00:02.975
      - 0.0
