+++ docs/_sources/auto_examples/example_sem_simulation.rst.txt
@@ -137,7 +137,7 @@ Three total crossover points of increasing tightness.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 6.517 seconds)
+   **Total running time of the script:** (0 minutes 6.411 seconds)
 
 
 .. _sphx_glr_download_auto_examples_example_sem_simulation.py:
