+++ docs/_sources/auto_examples/example_einzellens_deflection.rst.txt
@@ -75,7 +75,7 @@ between aperture_width and outer_diameter as well, for cleaner field configurati
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 2.879 seconds)
+   **Total running time of the script:** (0 minutes 2.975 seconds)
 
 
 .. _sphx_glr_download_auto_examples_example_einzellens_deflection.py:
