### 2025-05-14 #242 @rolypolytoy
@coatless,
I've made the following changes to the website (https://rolypolytoy.github.io/picht/):
Added a 'Getting Started' section that's first on the sidebar and provides explanatory content, explains the scope of the library and provides additional information (https://rolypolytoy.github.io/picht/gettingstarted.html).

Added a Computational Physics section that explains the formalism of the library to better understand the computational physics concepts it uses (implicitly functions as both an 'About' section and a 'Contributing' section by explaining its internals to make it easier to contribute out-of-the-box). It explains how it solves for the electric field and gives a bit of background about FDM and why we use it: https://rolypolytoy.github.io/picht/physics.html

Manually created API documentation rather than using Sphinx Autodoc, making the API documentation noticeably cleaner: https://rolypolytoy.github.io/picht/api.html

Renamed the Tutorials section to Gallery, but otherwise kept it identical or near-identical: https://rolypolytoy.github.io/picht/auto_examples/index.html

Let me know about any more changes that'll improve Picht's presentation/appeal and any other pre-review steps necessary
