+++ picht.egg-info/requires.txt
@@ -1,5 +1,9 @@
-numpy==1.26.4
 matplotlib
 scipy
 numba
 mendeleev
+pyamg
+joblib
+cadquery
+h5py
+numpy==1.26.4
