### 2025-05-16 af96e5e
v2.1.0- magnetic lenses. unit tested, ui support from the get-go, better documentation for website
