+++ picht.egg-info/SOURCES.txt
@@ -5,7 +5,8 @@ setup.cfg
 setup.py
 picht/__init__.py
 picht/core.py
-picht/example.py
+picht/electrostaticlens.py
+picht/magnetostaticlens.py
 picht.egg-info/PKG-INFO
 picht.egg-info/SOURCES.txt
 picht.egg-info/dependency_links.txt
