+++ examples/README.rst
@@ -4,4 +4,4 @@ Gallery
 This gallery contains examples demonstrating how to use Picht, moving up from basic cylindrical electrodes, to 
 understanding the properties of unipotential lenses, to finally being able to design and chain complex multi-lens systems.
 
-Begin at electrode creation, work your way through the unipotential lens series, and then try your hand at the SEM example.
+Begin at electric and magnetic lens creation, work your way through the unipotential lens series, and then try your hand at the SEM example.
