+++ docs/auto_examples/sg_execution_times.rst
@@ -1,0 +1,49 @@
+
+:orphan:
+
+.. _sphx_glr_auto_examples_sg_execution_times:
+
+
+Computation times
+=================
+**00:14.249** total execution time for 5 files **from auto_examples**:
+
+.. container::
+
+  .. raw:: html
+
+    <style scoped>
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
+    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
+    </style>
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
+    <script type="text/javascript" class="init">
+    $(document).ready( function () {
+        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
+    } );
+    </script>
+
+  .. list-table::
+   :header-rows: 1
+   :class: table table-striped sg-datatable
+
+   * - Example
+     - Time
+     - Mem (MB)
+   * - :ref:`sphx_glr_auto_examples_02_example_einzel_focusing.py` (``02_example_einzel_focusing.py``)
+     - 00:04.466
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_01_example_cylindrical_lens.py` (``01_example_cylindrical_lens.py``)
+     - 00:03.693
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_05_example_magnetic.py` (``05_example_magnetic.py``)
+     - 00:03.576
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_04_example_sem.py` (``04_example_sem.py``)
+     - 00:01.449
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_03_example_einzel.py` (``03_example_einzel.py``)
+     - 00:01.066
+     - 0.0

[…]
