### 2025-05-16 #242 @rolypolytoy
Hi, just commenting here for transparency:
I've added a new update to Picht (v2.1.0) that adds magnetic lenses as an option. It's a minor update, and I've updated the API documentation, docstrings, computational physics section to all reflect the change, added pytest coverage for it as well (all tests passing!) and added a new gallery example for it as well. As any new updates happen and I add new features (because I'm still actively maintaining/developing Picht) CI/CD and documentation will always keep pace.

I have a full changelog at: https://github.com/rolypolytoy/picht/releases
