+++ README.md
@@ -1,5 +1,5 @@
 # Picht
-An open-source electrodynamics and electron optics tool that's intuitive, performant, and physically accurate. Great for simulating the dynamics of electrons and ions through electrostatic and magnetostatic lenses of variable geometries. Calculates and visualizes electric and magnetic fields, and electrodynamics through those fields, using multigrid methods for FDM.
+Electron optics in electric and magnetic lenses. Uses the axisymmetric view, computes static E- and B-fields, enables the parameterization of ion and electron beams, and allows you to view their trajectories in the vicinity of various kinds of electromagnetic lenses, with physically accurate and relativistically corrected dynamics. Allows for custom mesh sizes using the finite difference method, and is incredibly performant due to integrations with PyAMG, Numba, and Joblib. Works on all operating systems, is unit-tested, works locally and on Jupyter Notebooks, and all of the examples in the gallery are with under 100 lines of code and generated in under a minute of time.  
 
 ## Installation
 ```bash
@@ -8,66 +8,22 @@ pip install picht
 [![PyPI version](https://img.shields.io/pypi/v/picht.svg)](https://pypi.org/project/picht/) ![tests](https://github.com/rolypolytoy/picht/actions/workflows/tests.yml/badge.svg) [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15376240.svg)](https://doi.org/10.5281/zenodo.15376240)
 
-## Documentation
+## Gallery
 
-You can initialize an einzel lens in the following manner:
+All API documentation, computational physics methods, and tutorials can be found at the official website: https://rolypolytoy.github.io/picht/. Look at CONTRIBUTING.MD if you want to begin adding code to the codebase, which is in picht/core.py. 
 
-```python
-import numpy as np
-from picht import IonOpticsSystem, ElectrodeConfig
-import matplotlib.pyplot as plt
+### Electric Lens
+Focusing electrons with a cylindrical electrostatic lens. Reference implementation [here](https://rolypolytoy.github.io/picht/auto_examples/01_example_cylindrical_lens.html)
+![cylindricallens](https://github.com/user-attachments/assets/3e3e00a7-009d-4cf6-a026-860ee545307e)
 
-system = IonOpticsSystem(nr=100, nz=600, axial_size=0.6, radial_size = 0.1) #all grid units are in mm.
+### Magnetic Lens
+Focusing electrons with a magnetic lens. Reference implementation [here](https://rolypolytoy.github.io/picht/auto_examples/05_example_magnetic.html)
+![ezgif-61b10d1bda4d4b](https://github.com/user-attachments/assets/333a8a55-f8b3-47ad-8dd7-268c9f8edb46)
 
-system.add_einzel_lens(
-    position=20.0,
-    width=60.0,
-    aperture_center=50.0,
-    aperture_width=48.0,
-    outer_diameter=50.0,
-    focus_voltage=-7000,
-    gap_size = 4
-)
-potential = system.solve_fields()
+### Einzel Lens
+Focusing electrons with three electrodes in a unipotential lens arrangement. Reference implementation [here](https://rolypolytoy.github.io/picht/auto_examples/02_example_einzel_focusing.html).
+![einzel](https://github.com/user-attachments/assets/44ba059c-c857-42fe-8d60-5b0ea997467f)
 
-trajectories = system.simulate_beam(
-    energy_eV= 10000,  
-    start_z=0,
-    r_range=(0.0499925, 0.0500075),
-    angle_range=(0, 0),
-    num_particles=6,
-    simulation_time=2e-9
-)
-
[…]
