+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "picht"
-version = "2.1.1"
+version = "2.1.2"
 description = "Electron and ion optics simulation using the Finite Difference Method (FDM)"
 readme = "README.md"
 requires-python = ">=3.6"
