+++ paper.md
@@ -22,28 +22,16 @@ Electron microscopes, focused ion beam systems, and mass spectrometers are speci
 
 # Statement of need
 
-'Picht' is a speedy, intuitive, and general library for electron optics. Very few non-commercial solutions exist in the field, and the existence of a tool in electron optics with enough physical realism, a low enough learning curve, and useful enough visualizations, was what gave rise to Picht. Instead of complex meshing, Picht uses a simple API with customizable sizes and shapes for cylindrical electrodes and einzel lenses, with custom voltages. Electron/ion beams can be initialized with parameterized sizes and dimensions, energies, initial divergences, and timescales of simulation. This parameterizable approach enables easy scripting of electron and ion optics, rapid prototyping of beam and lens configurations, and the ability to vary the degree of granularity of simulations. 
-
-For example, Picht enables a 68-LOC implementation of a full [Scanning Electron Microscope](https://github.com/rolypolytoy/picht/blob/main/examples/sem.py), with full handling of beam acceleration and divergence, Wehnelt cylinders, and an electrostatic condenser and objective lens, with fully realistic crossover and focusing behavior. Picht reduces the technical expertise required to design complex systems of electron optics, and thus increases the population that can prototype and design systems meant for electron microscopy, or more broadly for nanofabrication and nano-characterization. 
 
 # State of the Field
 
-Picht uses the Finite Difference Method to solve for the electric field- a method used at least as far back as 1989 to solve problems of this exact nature [@rouse1989three]. We use successive over-relaxation to solve systems of linear equations [@HADJIDIMOS2000177], which allows us to solve the Laplacian ∇²V = 0, followed by the electric field E = - ∇V. The package also uses Numba for speeding up the most compute-heavy operations, providing near-C speeds while remaining in the Python ecosystem and all the usability advantages that brings [@Oden2020]. This is in contrast to the boundary element method (BEM), which is used in many commercial and open-source solvers. We favor using the finite difference method due to us using Dirichlet instead of Neumann boundary conditions, due to the large amount of existing literature on finite difference method-based approximations to Dirichlet problems [@Huy1986]. 
+Picht uses the Finite Difference Method to solve for the electric and magnetic fields- a method used at least as far back as 1989 to solve problems of this exact nature [@rouse1989three]. We use multigrid methods to solve systems of linear equations, which allows us to solve the Laplacian for electric potential ∇²V = 0, followed by the electric field E = - ∇V, as well as solve the Poisson equation ∇²A = -μ₀μᵣJ to solve for vector potential and then B = ∇ × A to find the magnetic field [@Jackson1999]. This provides us with a discretized grid with electric and magnetic field values for tens of thousands to millions of individual points. The package uses Numba to speed up the most compute-heavy operations, providing near-C speeds while remaining in the Python ecosystem and all the usability advantages that brings [@Oden2020]. This is in contrast to the boundary element method (BEM), which is used in many commercial and open-source solvers. We favor using the finite difference method due to us using Dirichlet instead of Neumann boundary conditions, due to the large amount of existing literature on finite difference method-based approximations to Dirichlet problems, enabling a more consistently accurate performance with the current state of knowledge in the field of computational physics [@Huy1986]. 
 
 We also, when calculating velocity, or acceleration, use relativistically-accurate formulae by computing the Lorentz factor
 $$\gamma = \frac{1}{\sqrt{1 - \frac{v^2}{c^2}}}$$
[…]
