+++ paper.bib
@@ -1,3 +1,21 @@
+@book{Taylor2000,
+  title={Special Relativity: A Modern Introduction},
+  author={Taylor, Edwin F. and Wheeler, John Archibald},
+  year={2000},
+  publisher={W. H. Freeman},
+  isbn={978-0716723271}
+}
+
+@book{Jackson1999,
+  title={Classical Electrodynamics},
+  author={Jackson, John David},
+  year={1999},
+  edition={3},
+  publisher={John Wiley \& Sons},
+  address={New York},
+  isbn={978-0471309321}
+}
+
 @article{rouse1989three,
   title={Three-dimensional computer modeling of electrostatic and magnetic electron optical components},
   author={Rouse, John and Munro, Eric},
