+++ paper.md
@@ -20,18 +20,21 @@ bibliography: paper.bib
 
 Electron microscopes, focused ion beam systems, and mass spectrometers are specialized devices important to characterizing and fabricating nanotechnology. Electron optics is the field that deals with designing and shaping beams of charged particles, and relies heavily on computational methods. However, many important kinds of electrostatic lenses don't have easy analytical solutions [@hawkes_principles_1989], which increases the reliance of researchers on either experiment or simulation, when designing and understanding such systems. Numerical integration solvers are suitable for problems which require solving differential equations, and the backwards differentiation formula (BDF) is a suitable solver for use cases with potentially unstable numerical results [@Cash2015]. The package first calculates the electric field generated by the initialized electrodes or electrostatic lenses. It then calculates the particle trajectories based on the specified beam characteristics, and provides a visualization of individual particle trajectories across the specified timescale. It also fully calculates relativistic effects, and so accurately simulates particle dynamics at the keV and MeV scale, and assumes grounded boundaries, or Dirichlet boundary conditions. 
 
-# Statement of need
-
-
-# State of the Field
-
 Picht uses the Finite Difference Method to solve for the electric and magnetic fields- a method used at least as far back as 1989 to solve problems of this exact nature [@rouse1989three]. We use multigrid methods to solve systems of linear equations, which allows us to solve the Laplacian for electric potential ∇²V = 0, followed by the electric field E = - ∇V, as well as solve the Poisson equation ∇²A = -μ₀μᵣJ to solve for vector potential and then B = ∇ × A to find the magnetic field [@Jackson1999]. This provides us with a discretized grid with electric and magnetic field values for tens of thousands to millions of individual points. The package uses Numba to speed up the most compute-heavy operations, providing near-C speeds while remaining in the Python ecosystem and all the usability advantages that brings [@Oden2020]. This is in contrast to the boundary element method (BEM), which is used in many commercial and open-source solvers. We favor using the finite difference method due to us using Dirichlet instead of Neumann boundary conditions, due to the large amount of existing literature on finite difference method-based approximations to Dirichlet problems, enabling a more consistently accurate performance with the current state of knowledge in the field of computational physics [@Huy1986]. 
 
 We also, when calculating velocity, or acceleration, use relativistically-accurate formulae by computing the Lorentz factor
 $$\gamma = \frac{1}{\sqrt{1 - \frac{v^2}{c^2}}}$$
 We then use the energy-momentum formalism of special relativity to calculate radial and axial velocities and forces from positions and momenta using numerical differentiation with respect to time [@Taylor2000]. This, coupled with an ordinary differential equation solver using the backwards differentiation method using an integration with scipy, which we've found provides more consistently stable results than Runge-Kutta methods applied with the same parameters [@Cash2015]. 
 
-Picht can generate electron and ion beams of varying sizes- from filling the entire space radially, to generating a beam of electrons microns wide, to enable beams on the order of magnitude of those produced from a tungsten (W) source [@LI2023111837]. Additional parameterizations Picht enables includes a parameterization of the geometry of electrodes, the 
+Picht can generate electron and ion beams of varying sizes- from filling the entire space radially, to generating a beam of electrons microns wide, to enable beams on the order of magnitude of those produced from a tungsten (W) source [@LI2023111837]. Additional parameterizations Picht enables includes a parameterization of the geometry of electrodes and the voltage this is held at, the ability to parameterize magnetomotive force and relative permeability, and the ability to toggle viewing the electric and magnetic fields, to provide a user-friendly UI to interface with. 
+
+
+# Statement of need
+
+
+
+# State of the Field
+
 
 
 # References
