+++ paper.bib
@@ -1,104 +1,1 @@
-@book{Taylor2000,
-  title={Special Relativity: A Modern Introduction},
-  author={Taylor, Edwin F. and Wheeler, John Archibald},
-  year={2000},
-  publisher={W. H. Freeman},
-  isbn={978-0716723271}
-}
 
-@book{Jackson1999,
-  title={Classical Electrodynamics},
-  author={Jackson, John David},
-  year={1999},
-  edition={3},
-  publisher={John Wiley \& Sons},
-  address={New York},
-  isbn={978-0471309321}
-}
-
-@article{rouse1989three,
-  title={Three-dimensional computer modeling of electrostatic and magnetic electron optical components},
-  author={Rouse, John and Munro, Eric},
-  journal={Journal of Vacuum Science \& Technology B: Microelectronics Processing and Phenomena},
-  volume={7},
-  number={6},
-  pages={1891--1897},
-  year={1989},
-  publisher={American Vacuum Society}
-}
-
-@article{HADJIDIMOS2000177,
-title = {Successive overrelaxation (SOR) and related methods},
-journal = {Journal of Computational and Applied Mathematics},
-volume = {123},
-number = {1},
-pages = {177-199},
-year = {2000},
-note = {Numerical Analysis 2000. Vol. III: Linear Algebra},
-issn = {0377-0427},
-doi = {https://doi.org/10.1016/S0377-0427(00)00403-9},
-url = {https://www.sciencedirect.com/science/article/pii/S0377042700004039},
-author = {A. Hadjidimos},
-keywords = {Iterative methods, Jacobi method, Gauss–Seidel method, SOR methods},
-abstract = {Covering the last half of the 20th century, we present some of the basic and well-known results for the SOR theory and related methods as well as some that are not as well known. Most of the earlier results can be found in the excellent books by Varga (Matrix Iterative Analysis, Prentice-Hall, Englewood Cliffs, NJ, 1962) Young (Iterative Solution of Large Linear systems, Academic Press, New York, 1971) and Berman and Plemmons (Nonnegative Matrices in the Mathematical Sciences, SIAM, Philadelphia, PA, 1994) while some of the most recent ones are given in the bibliography of this paper. In this survey, both the point and the block SOR methods are considered for the solution of a linear system of the form Ax=b, where A∈Cn,n and b∈Cn⧹{0}. Some general results concerning the SOR and related methods are given and also some more specific ones in cases where A happens to possess some further property, e.g., positive definiteness, L-, M-, H-matrix property, p-cyclic consistently ordered property etc.}
-}
-
-@book{hawkes_principles_1989,
-	title = {Principles of {Electron} {Optics}: {Applied} geometrical optics},
-	isbn = {978-0-12-333352-0},
-	shorttitle = {Principles of {Electron} {Optics}},
-	publisher = {Academic Press},
-	author = {Hawkes, P. W. and Kasper, E.},
-	year = {1989},
-}
-
-@Inbook{Cash2015,
-author="Cash, Jeff R.",
-editor="Engquist, Bj{\"o}rn",
-title="Backward Differentiation Formulae",
-bookTitle="Encyclopedia of Applied and Computational Mathematics",
-year="2015",
-publisher="Springer Berlin Heidelberg",
-address="Berlin, Heidelberg",
-pages="97--101",
-isbn="978-3-540-70529-1",
-doi="10.1007/978-3-540-70529-1_94",
-url="https://doi.org/10.1007/978-3-540-70529-1_94"
-}
-
-@INPROCEEDINGS{9092407,
-  author={Oden, Lena},
-  booktitle={2020 28th Euromicro International Conference on Parallel, Distributed and Network-Based Processing (PDP)}, 
-  title={Lessons learned from comparing C-CUDA and Python-Numba for GPU-Computing}, 
[…]
