+++ paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-Electron optics is the field of physics that deals with the study of electron and ion beams under the influence of electric and magnetic 'lenses'. Electromagnetic lenses are called 'lenses' because they make charged particle beams converge, diverge and reflect, in a way analogous to how optical lenses affect beams of light. Simulating and designing systems of electron optics are important to modern materials science and nanotechnology, since the design of many integral components used in their research, like electron microscopes, ion implantation and milling machines, and mass spectrometers. Picht provides researchers and engineers with the ability to design and simulate electron optics to reduce the barrier of entry of research in this field. 
+Electron optics is the field of physics that deals with the study of electron and ion beams under the influence of electric and magnetic 'lenses'. Electromagnetic lenses are called 'lenses' because they make charged particle beams converge, diverge and reflect, in a way analogous to how optical lenses affect beams of light. Simulating and designing systems of electron optics are important to modern materials science and nanotechnology, since the design of many integral components used in their research, like electron microscopes, ion implantation and milling machines, and mass spectrometers. Picht provides researchers and engineers with the ability to design and simulate such systems to reduce the barrier of entry of research in this field. 
 
 # Statement of Need
 
