+++ paper.md
@@ -23,7 +23,9 @@ Electron optics is the field of physics that deals with the study of electron an
 
 # Statement of Need
 
+`Picht` is a Python package to simulate electron optics systems. Using Python enables an easy and intuitive API, without sacrificing on performance when compared to writing the package in a compiled language (like C++). This is because it uses libraries like Numba [lam2015numba], which provide performance in some cases comparable to CUDA while maintaining the standards of usability expected of libraries in the Python ecosystem  [@9092407]. `Picht` exists to provide an open-source alternative to commercial electrodynamics tools like SIMION and COMSOL, without sacrifices in user-friendliness or performance. It solves magnetic fields using an optimized multigrid solver by PyAMG [pyamg2023], and uses joblib [joblib2020] for parallelizing trajectory calculations. 
 
+`Picht` is architecturally focused on parameterization rather than scripting. The classes are constrained, and creating a beam is simple, however adjusting parameters massively affects trajectories and outcomes. The intended workflow is creating a system, observing the user interface, and tweaking voltage, magnetomotive force, permeability, or geometric features, until the system fits its intended purpose. It thus enables research or education in electron optics for amateurs and experts alike, due to its intentionally constrained syntax, but computationally efficient physics engine.
 
 # Computational Physics
 
