+++ paper.md
@@ -29,6 +29,36 @@ Electron optics is the field of physics that deals with the study of electron an
 
 # Computational Physics
 
+The main equation `Picht` uses to solve for its magnetic field is Poisson's equation:
 
+$$\nabla^2 A = -\mu_0 \mu_r J$$
+
+Where $\nabla^2$ is the Laplacian operator, A is the vector potential, $\mu_0$ is the permeability of free space, $\mu_r$ is the dimensionless value for relative permeability of the material, and J is the magnetomotive force in ampere-turns [jackson1999classical]. We use Dirichlet masking to specify magnetomotive force values, as well as relative permeability values, where the user specifies a magnetic lens is. Then, PyAMG solves for vector potential based on the resolution of the grid the user specified.
+
+From this, we can use the equation: 
+
+$$B = \nabla \times A$$
+
+To compute the magnetic field 'B' from the curl of the vector potential A. We use the axisymmetric formulation for curl, by making the assumption that the system is cylindrically symmetric around its optical axis, which is a reasonable assumption to make in electron optics [hawkes1972electron].
+
+To compute the electric field, we solve Laplace's equation first:
+
+$$\nabla^2 V = 0$$
+
+Where V is electric potential  [jackson1999classical]. We use Dirichlet masking to set the exact voltage amount the user specified, in the selected geometry of the lens, and then the above equation is computed using numerical methods. From the electric potential field, the electric field can be calculated through the following equation: 
+
+$$E = -\nabla V$$
+
+Given the electric and magnetic field components at every point, particle trajectories can be computed. For electric effects, we use the Lorentz equation for electrostatics:
+
+$$F = qE$$
+
+Where q is charge and E is the electric field component. However, for the magnetic field, we don't use the Lorentz equation, because the full Lorentz treatment magnetic focusing is prohibitively expensive to simulate. To provide accuracy and efficiency, we instead use the paraxial ray equation for magnetic lenses, which is a reasonable approximation at small angles from the axis [hawkes1972electron]. The form we use in our code is:
+
+$$F_r =  -\frac{q^2 B_z^2}{4m} r$$
+
+Where $F_r$ is the radial component of force, q is the charge of the particle, $B_z$ is the axial component of the magnetic field, m is the mass of the particle, and r is the distance of the particle from the central axis.
+
+Combining this treatment of electric and magnetic components of force enables us to simulate systems with both kinds of lenses in an accurate and performant manner.
 
 # References
