+++ paper.bib
@@ -1,1 +1,57 @@
+@INPROCEEDINGS{9092407,
+  author={Oden, Lena},
+  booktitle={2020 28th Euromicro International Conference on Parallel, Distributed and Network-Based Processing (PDP)}, 
+  title={Lessons learned from comparing C-CUDA and Python-Numba for GPU-Computing}, 
+  year={2020},
+  volume={},
+  number={},
+  pages={216-223},
+  keywords={Python;Graphics processing units;Libraries;Kernel;Indexes;Benchmark testing},
+  doi={10.1109/PDP50117.2020.00041}
+}
 
+@article{pyamg2023,
+  author    = {Nathan Bell and Luke N. Olson and Jacob Schroder and Ben Southworth},
+  title     = {{PyAMG}: Algebraic Multigrid Solvers in Python},
+  journal   = {Journal of Open Source Software},
+  year      = {2023},
+  publisher = {The Open Journal},
+  volume    = {8},
+  number    = {87},
+  pages     = {5495},
+  doi       = {10.21105/joss.05495},
+  url       = {https://joss.theoj.org/papers/10.21105/joss.05495}
+}
+
+@inproceedings{lam2015numba,
+  title={Numba: A llvm-based python jit compiler},
+  author={Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
+  booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
+  pages={1--6},
+  year={2015}
+}
+
+@software{joblib2020,
+  title = {Joblib: running Python functions as pipeline jobs},
+  author = {{Joblib Development Team}},
+  year = {2020},
+  url = {https://joblib.readthedocs.io/}
+}
+
+@book{jackson1999classical,
+  title={Classical Electrodynamics},
+  author={Jackson, John David},
+  year={1999},
+  edition={3},
+  publisher={Wiley},
+  address={New York, NY},
+  isbn={9780471309321}
+}
+
+@book{hawkes1972electron,
+  title={Electron Optics and Electron Microscopy},
+  author={Hawkes, P. W.},
+  year={1972},
+  publisher={Taylor \& Francis},
+  address={London}
+}
