### 2025-05-17 e7d7a85
Update magnetostaticlens.py
