+++ picht/core.py
@@ -38,7 +38,7 @@ class MagneticLensConfig:
 class MagneticField:
    """
    This class handles everything related to the magnetic vector potential field, including
-   adding magnetic lenses, and solving for the magnetic field. Analagous to PotentialField
+   adding magnetic lenses, and solving for the magnetic field. Analagous to ElectricField
    but solves the magnetic field instead of the electric field.
    
    Attributes:
@@ -83,10 +83,10 @@ class MagneticField:
        self.Br = np.zeros((self.nz, self.nr))
        self.lens_config = None
       
-   def add_magnetic_lens(self, config: MagneticLensConfig):
+   def add_magnetic_lens(self, config):
        """
        Adds a magnetic lens to the field and handles all necessary calculations.
-       Analagous to add_electrode() in PotentialField.
+       Analagous to add_electrode() in ElectricField.
 
        Parameters:
            config (MagneticLensConfig): Configuration parameters for the magnetic lens.
@@ -121,7 +121,7 @@ class MagneticField:
        Builds a sparse matrix for the Laplacian ∇²A = -μ₀μᵣJ, 
        and implements Neumann boundary conditions at all boundaries, to simulate
        how magnetic fields behave at metal boundaries. Identical implementation
-       details as build_laplacian_matrix() in PotentialField.
+       details as build_laplacian_matrix() in ElectricField.
 
        Parameters:
            mask (ndarray): Boolean array, true where magnetic materials exist, false elsewhere.
@@ -140,6 +140,11 @@ class MagneticField:
        def idx(i, j):
            return i * self.nr + j
 
+       def harmonic_mean(a, b):
+           if a == 0 or b == 0:
+               return 0
+           return 2 * a * b / (a + b)
+
        ap_center = (self.lens_config.ap_start + self.lens_config.ap_width / 2) if self.lens_config else 0
        epsilon = 1e-10
 
@@ -147,49 +152,85 @@ class MagneticField:
            for j in range(self.nr):
                k = idx(i, j)
                
-               if i == 0:
+               if i < 2 or i >= self.nz - 2:
                    A[k, k] = -1.0
-                   A[k, idx(i+1, j)] = 1.0
+                   if i == 0:
+                       A[k, idx(min(i+1, self.nz-1), j)] = 1.0
+                   elif i == 1:
+                       A[k, idx(min(i+1, self.nz-1), j)] = 1.0
+                   elif i == self.nz - 2:
+                       A[k, idx(max(i-1, 0), j)] = 1.0
+                   else:
+                       A[k, idx(max(i-1, 0), j)] = 1.0
                    b[k] = 0.0
-               elif i == self.nz - 1:
+               elif j < 2 or j >= self.nr - 2:
                    A[k, k] = -1.0
-                   A[k, idx(i-1, j)] = 1.0
-                   b[k] = 0.0
-               elif j == 0:
-                   A[k, k] = -1.0
-                   A[k, idx(i, j+1)] = 1.0
-                   b[k] = 0.0
-               elif j == self.nr - 1:
-                   A[k, k] = -1.0
-                   A[k, idx(i, j-1)] = 1.0
+                   if j == 0:
+                       A[k, idx(i, min(j+1, self.nr-1))] = 1.0
+                   elif j == 1:
+                       A[k, idx(i, min(j+1, self.nr-1))] = 1.0
+                   elif j == self.nr - 2:
+                       A[k, idx(i, max(j-1, 0))] = 1.0
+                   else:
+                       A[k, idx(i, max(j-1, 0))] = 1.0
                    b[k] = 0.0
                else:
                    mu_center = self.mu_r[i, j]
[…]
