+++ picht/__init__.py
@@ -1,7 +1,7 @@
-__version__ = "2.1.0"
+__version__ = "2.1.3"
 
 from .core import (
-    PotentialField,
+    ElectricField,
     ParticleTracer,
     ElectrodeConfig,
     EinzelLens,
