### 2025-05-18 72bb413
v2.1.3- more accurate magnetic field meshing
