+++ picht.egg-info/SOURCES.txt
@@ -5,7 +5,7 @@ setup.cfg
 setup.py
 picht/__init__.py
 picht/core.py
-picht/magnetostaticlens.py
+picht/sem.py
 picht.egg-info/PKG-INFO
 picht.egg-info/SOURCES.txt
 picht.egg-info/dependency_links.txt
