### 2025-05-18 46c7b03
2.1.6- electronoptics and electricfield, neumann boundaries
