+++ picht/core.py
@@ -13,6 +13,305 @@ from joblib import Parallel, delayed
 import multiprocessing
 import os
 
+@dataclass
+class MagneticLensConfig:
+    """
+    Parameterizes a simple magnetic lens.
+    
+    Attributes:
+        start (float): Position where the magnetic lens begins on the z-axis in grid units.
+        length (float): Length of the magnetic lens on the z-axis in grid units.
+        ap_start (float): Starting position of the aperture on the r-axis in grid units.
+        ap_width (float): Width of the aperture in the lens on the r-axis in grid units.
+        outer_diameter (float): Diameter of the magnetic lens on the r-axis in grid units.
+        mu_r (float): Relative magnetic permeability of the lens material in dimensionless units.
+        mmf (float): Magnetomotive force of the lens in ampere-turns.
+    """
+    start: float
+    length: float
+    ap_start: float
+    ap_width: float
+    outer_diameter: float
+    mu_r: float
+    mmf: float
+
+class MagneticField:
+   """
+   This class handles everything related to the magnetic vector potential field, including
+   adding magnetic lenses, and solving for the magnetic field. Analagous to ElectricField
+   but solves the magnetic field instead of the electric field.
+   
+   Attributes:
+       nz (int): Number of grid points in the z-axis.
+       nr (int): Number of grid points in the r-axis.
+       axial_size (float): Size of the z-axis in meters.
+       radial_size (float): Size of the r-axis in meters.
+       dz (float): Grid spacing in z-axis per z-unit, in meters, found by axial_size/nz.
+       dr (float): Grid spacing in r-axis per r-unit, in meters, found by radial_size/nr.
+       vector_potential (ndarray): 2D array with magnetic vector potential values in Tesla-meters.
+       magnetic_mask (ndarray): Boolean mask for magnetic material positions.
+       mu_r (ndarray): 2D array of relative permeability values.
+       current_density (ndarray): 2D array of current density values in amperes per square meter.
+       Bz (ndarray): Z-component of magnetic field in Tesla.
+       Br (ndarray): R-component of magnetic field in Tesla.
+       lens_config (MagneticLensConfig): Configuration of the magnetic lens.
+   """
+  
+   def __init__(self, electron_optics):
+       """
+       Initializes the magnetic field with the specific dimensions. Generally when
+       calculating magnetic lenses, you need to have finer meshes than with solely
+       electrostatic lenses, because there are more discretization artefacts
+       because of the paraxial ray equation. Since we use multigrid methods 
+       this isn't exceptionally costly.
+       
+       Parameters:
+           electron_optics (ElectronOptics): The parent ion optics system containing
+               field dimensions and grid parameters.
+       """
+       self.nz = electron_optics.field.nz
+       self.nr = electron_optics.field.nr
+       self.axial_size = electron_optics.field.axial_size
+       self.radial_size = electron_optics.field.radial_size
+       self.dz = electron_optics.field.dz
+       self.dr = electron_optics.field.dr
+       self.vector_potential = np.zeros((self.nz, self.nr))
+       self.magnetic_mask = np.zeros((self.nz, self.nr), dtype=bool)
+       self.mu_r = np.ones((self.nz, self.nr))
+       self.current_density = np.zeros((self.nz, self.nr))
+       self.Bz = np.zeros((self.nz, self.nr))
+       self.Br = np.zeros((self.nz, self.nr))
+       self.lens_config = None
+      
+   def add_magnetic_lens(self, config: MagneticLensConfig):
+       """
+       Adds a magnetic lens to the field and handles all necessary calculations.
[…]
