### 2025-05-19 17650b4
2.1.7: bugfix, clamps nans and infs in ode solver
