+++ picht/sem.py
@@ -1,5 +1,5 @@
 import numpy as np
-from core import ElectronOptics, ElectrodeConfig, MagneticLensConfig, Export
+from picht import ElectronOptics, ElectrodeConfig, MagneticLensConfig, Export
 import matplotlib.pyplot as plt
 
 system = ElectronOptics(nr=1000, nz=4000, axial_size=0.4, radial_size = 0.01)
@@ -42,7 +42,6 @@ cathode = ElectrodeConfig(
 system.add_electrode(anode)
 system.add_electrode(cathode)
 #Electron Gun- Finished.
-
 
 system.solve_fields()
 
