+++ picht/core.py
@@ -35,6 +35,12 @@ class MagneticLensConfig:
     mu_r: float
     mmf: float
 
+def sanitize_field_values(field_array, max_value=1e12):
+    field_array = np.nan_to_num(field_array, nan=0.0)
+    field_array = np.nan_to_num(field_array, posinf=max_value, neginf=-max_value)
+    field_array = np.clip(field_array, -max_value, max_value)
+    return field_array
+
 class MagneticField:
    """
    This class handles everything related to the magnetic vector potential field, including
@@ -215,21 +221,36 @@ class MagneticField:
        """
        A, b = self.build_laplacian_matrix(self.magnetic_mask, self.vector_potential)
        
+       if np.any(np.isnan(A.data)) or np.any(np.isinf(A.data)):
+           A.data = sanitize_field_values(A.data, 1e12)
+       
+       if np.any(np.isnan(b)) or np.any(np.isinf(b)):
+           b = sanitize_field_values(b, 1e12)
+       
        scale_factor = 1.0 / max(np.max(np.abs(A.data)), 1e-10)
        A_scaled = A * scale_factor
        b_scaled = b * scale_factor
        
-       ml = pyamg.smoothed_aggregation_solver(
-           A_scaled, 
-           max_coarse=10,
-           strength='symmetric',
-           smooth='jacobi',
-           improve_candidates=None
-       )
+       A_scaled.data = sanitize_field_values(A_scaled.data, 1e12)
+       b_scaled = sanitize_field_values(b_scaled, 1e12)
        
-       x0 = np.zeros_like(b_scaled)
-       x = ml.solve(b_scaled, x0=x0, tol=convergence_threshold, maxiter=int(max_iterations))
-       self.vector_potential = x.reshape((self.nz, self.nr))
+       try:
+           ml = pyamg.smoothed_aggregation_solver(
+               A_scaled, 
+               max_coarse=10,
+               strength='symmetric',
+               smooth='jacobi',
+               improve_candidates=None
+           )
+           
+           x0 = np.zeros_like(b_scaled)
+           x = ml.solve(b_scaled, x0=x0, tol=convergence_threshold, maxiter=int(max_iterations))
+           
+           x = sanitize_field_values(x, 1e12)
+           self.vector_potential = x.reshape((self.nz, self.nr))
+           
+       except Exception as e:
+           self.vector_potential = sanitize_field_values(self.vector_potential, 1e12)
        
        self.calculate_b_from_a()
        
@@ -242,6 +263,7 @@ class MagneticField:
        """
        ap_center = (self.lens_config.ap_start + self.lens_config.ap_width / 2) if self.lens_config else 0
        epsilon = 1e-10
+       max_field = 1e6
        
        for i in range(self.nz):
            for j in range(self.nr):
@@ -273,6 +295,10 @@ class MagneticField:
                    dA_dr = -dA_dr
                
                self.Bz[i, j] = dA_dr + self.vector_potential[i, j] / r_dist
+       
+       self.Bz = sanitize_field_values(self.Bz, max_field)
+       self.Br = sanitize_field_values(self.Br, max_field)
+       self.vector_potential = sanitize_field_values(self.vector_potential, max_field)
   
    def get_field_at_position(self, z: float, r: float) -> Tuple[float, float]:
        """
@@ -288,10 +314,18 @@ class MagneticField:
        if 0 <= z < self.axial_size and 0 <= r < self.radial_size:
            i = int(min(max(0, z / self.dz), self.nz - 1))
            j = int(min(max(0, r / self.dr), self.nr - 1))
-           return self.Bz[i, j], self.Br[i, j]
[…]
