### 2025-05-19 5028ccf
better NaN and inf handling, fixed radians
