+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "picht"
-version = "2.1.7"
+version = "2.1.8"
 description = "Electron optics simulation tool"
 readme = "README.md"
 requires-python = ">=3.6"
