### 2025-05-19 6ef5a88
v2.2.0-pre
fixed the center discretization bug by switching A_phi bcs to dirichlet over neumann.
