+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "picht"
-version = "2.1.8"
+version = "2.2.0"
 description = "Electron optics simulation tool"
 readme = "README.md"
 requires-python = ">=3.6"
