### 2025-06-30 #242 @coatless
@rolypolytoy just an update, we've found an editor for your review!

I'm happy to appoint @HaoZeke (Rohit Goswami) as the editor. I'll let him introduce himself.

(P.S He's recovering from an illness. So, the intro may not be immediate.)
