### 2025-07-09 #242 @rolypolytoy
Hi @HaoZeke. I'm looking forward to getting any and all feedback that can help me improve Picht. Feel free to take your time.
